% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cft.R
\name{check.cft}
\alias{check.cft}
\title{Check that counterfactual decreases exposure levels}
\usage{
check.cft(cft, X)
}
\arguments{
\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for 
the Population Attributable Fraction \code{\link{paf}} where counterfactual
is 0 exposure.}

\item{X}{Random sample (vector or matrix) which includes exposure and
covariates. or sample mean.}
}
\description{
Function that checks whether the counterfactual function of \code{\link{pif}} decreases the exposure values
}
\examples{
#Example 1 
cft <- function(X){0.5*X}
X   <- runif(100, 0,2)
check.cft(cft, X)


}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}

