% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_confidence.R
\name{check.confidence}
\alias{check.confidence}
\title{Check confidence is between zero and less than a hundred}
\usage{
check.confidence(confidence)
}
\arguments{
\item{confidence}{Confidence level desired}
}
\value{
TRUE if Confidence is well defined between zero and a hundred
}
\description{
Function that verifies confidence level selected is between zero and less than a hundred
}
\examples{
#Example 1 
confidence <- 95
check.confidence(confidence)


}

