% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.exposure.R
\name{check.exposure}
\alias{check.exposure}
\title{Check Exposure values}
\usage{
check.exposure(X)
}
\arguments{
\item{X}{Matrix of exposure that is evaluated in rr}
}
\value{
TRUE if all exposure values are greater or equal to zero
}
\description{
Function that verifies if all exposure values are greater or equal to zero (exposure values can not be negative)
}
\examples{
#Example 1 
X <- runif(500, 0,1)
check.exposure(X)


}

