% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_integrals.R
\name{check.integrals}
\alias{check.integrals}
\title{Check Integrals}
\usage{
check.integrals(meanrr, meancft)
}
\arguments{
\item{meanrr}{Mean of relative risk without counterfactual}

\item{meancft}{Mean of relative risk with counterfactual}
}
\value{
bool   TRUE if as desired
}
\description{
Function for checking that the integrals of \code{\link{pif}} make sense.
}
\examples{
check.integrals(1,0)

\dontrun{
check.integrals(0,1)
check.integrals(1,-1)
}

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}

