% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_read_csv.R
\name{check.read.csv}
\alias{check.read.csv}
\title{Check csv file was correctly defined in R shiny input}
\usage{
check.read.csv(file)
}
\arguments{
\item{file}{Contains file information from "fileInput"}
}
\description{
Function that verifies the input is not empty, and that the input is in a csv format
}
\author{
Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com}

Dalia Camacho-García-Formentí \email{daliaf172@gmail.com}
}

