% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_read_Rfun.R
\name{check.read.Rfun}
\alias{check.read.Rfun}
\title{Check an R function has been given with a specific name}
\usage{
check.read.Rfun(file, fname)
}
\arguments{
\item{file}{Contains file information from "fileInput"}

\item{fname}{Name of the function as a string}
}
\description{
Function that verifies the input is not empty. Verifies the input is an R file. Verifies the R file contained a function with a specific name. If everything is correct returns TRUE, if not a warning label.
}
\author{
Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com}

Dalia Camacho-García-Formentí \email{daliaf172@gmail.com}
}

