% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rr.R
\name{check.rr}
\alias{check.rr}
\title{Check Relative Risk}
\usage{
check.rr(X, thetahat, rr, tol = 1e-08)
}
\arguments{
\item{X}{Matrix of exposure that is evaluated in rr}

\item{thetahat}{Point estimate of theta}

\item{rr}{Relative risk function}

\item{tol}{Tolerance for equality}
}
\value{
TRUE if relative risk is as desired
}
\description{
Function for checking that relative risk equals 1 when evaluated in 0
}
\examples{
check.rr(as.matrix(rnorm(100)), 1, function(X, theta){exp(X*theta)})

}

