% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_thetas.R
\name{check.thetas}
\alias{check.thetas}
\title{Check thetas against method}
\usage{
check.thetas(thetavar, thetahat, thetalow, thetaup, method)
}
\arguments{
\item{thetavar}{Variance of theta}

\item{thetahat}{Point estimate of theta}

\item{thetalow}{Lower bound of theta's CI}

\item{thetaup}{Upper bound of theta's CI}

\item{method}{Method of CI's}
}
\value{
bool     Boolean variable indicating if hypothesis are matched
}
\description{
Function for checking that the thetas are .correctly specified according to chosen method
}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}

