% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Xvar.R
\name{check.xvar}
\alias{check.xvar}
\title{Check Xvar was defined}
\usage{
check.xvar(Xvar)
}
\arguments{
\item{Xvar}{Input of variance of X}
}
\value{
Xvar
}
\description{
Function that verifies Xvar was defined
}
\examples{
#Example 1 
Xvar <- NA
check.xvar(Xvar)

#Example 2
Xvar <- 0.2
check.xvar(Xvar)

}

