% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{coef.pif_sim}
\alias{coef.pif_sim}
\title{Extract point estimate of potential impact fraction}
\usage{
\method{coef}{pif_sim}(x, ...)
}
\arguments{
\item{...}{additional arguments}

\item{pif_sim}{A \code{pif_sim} object}
}
\description{
Method for obtaining the point estimate of a \code{pif_sim} object (i.e. the
point-estimate of the potential impact fraction or population attributable fraction via
the bootstrap).
}
\examples{
some_disease <- pif_sim(pif_simulations = rnorm(100))
coef(some_disease)

}
