% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{confint.pif_class}
\alias{confint.pif_class}
\title{Extract confidence intervals for the potential impact fraction or the
population attributable fraction}
\usage{
\method{confint}{pif_class}(x, parm = NULL, level = 0.95, method = c("wald", "percentile"), ...)
}
\arguments{
\item{x}{A \code{pif_class} object}

\item{parm}{a specification of which parameters are to be given confidence intervals,
either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{(\code{double}) Confidence level for the uncertainty interval.
Defaults to \code{0.95}.}

\item{method}{(\code{string}) either \code{'wald'} (recommended) or \code{'percentile'} for
uncertainty intervals based upon the bootstrap's percentile or the Wald-type approximation
(see confidence interval section)}

\item{...}{additional methods to pass to \code{\link[=mean]{mean()}} and \code{\link[stats:sd]{stats::sd()}}
if \code{wald} or \code{\link[stats:quantile]{stats::quantile()}} if \code{percentile} type intervals}
}
\description{
Method for obtaining confidence intervals for a \code{pif_class} object
}
\section{Confidence Intervals}{


Confidence intervals are estimated via the two methods specified
by \insertCite{arnab2017survey;textual}{pifpaf}.
\enumerate{
\item \strong{Wald-type confidence intervals} (which are more precise) are of the form:
\mjdeqn{
\widehat{\text{PIF}} \pm t_{1 - \alpha/2}
\sqrt{\widehat{\text{Var}}_{\text{B}}\big[\widehat{\text{PIF}}\big]}
}{pif +- qt(1 - alpha/2)*sqrt(var(pif))}
where \mjseqn{t_{1 - \alpha/2}} stands for the percent points at level \mjseqn{1 - \alpha/2}
of Student's t-distribution, and
\mjseqn{\widehat{\text{Var}}_{\text{B}}\big[\widehat{\text{PIF}}\big]}
represents the estimated variance (via bootstrap) of the potential impact fraction estimator.
\item \strong{Percentile confidence intervals} (less precise) are of the form:
\mjdeqn{
\Big[\widehat{\text{PIF}}_{\text{B},\alpha/2}, \widehat{\text{PIF}}_{\text{B},1-\alpha/2}\Big]
}{quantile(pif, c(alpha/2, 1-alpha/2))}
where \mjseqn{\widehat{\text{PIF}}_{\text{B},k}} represents the kth sample quantile
from the bootstrap simulation of the potential impact fraction estimator.
}
}

\examples{
#Example 1
data(ensanut)
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pifsim <- pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)
confint(pifsim)
}
