% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensanut.R
\docType{data}
\name{ensanut}
\alias{ensanut}
\title{ENSANUT 2018}
\format{
Information contains: A tibble containing the following information
\itemize{
\item \code{age} Age of the individual (years).
\item \code{sex} Sex of the individual (\code{Male} or \code{Female}).
\item \code{weight} Survey weights associated to the survey.
\item \code{strata} Strata of the survey.
\item \code{hypertension} If individual has hypertension \code{TRUE}, \code{FALSE} otherwise.
\item \code{delta_na_phase_1} Proposed change in sodium consumption (counterfactual).
\item \code{systolic_blood_pressure} Mean systolic blood pressure of the individual registered by survey.
\item \code{age_group} Age group of the individual
}
}
\usage{
ensanut
}
\description{
Data on the Mexican \href{https://ensanut.insp.mx/}{National Health and Nutrition Survey of 2018 (ENSANUT-2018)}.
}
\examples{
ensanut
# Create a survey design as follows:
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
}
\keyword{datasets}
