% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntryMult.R
\name{EntryMult}
\alias{EntryMult}
\title{Sum of entry multiplication}
\usage{
EntryMult(X, Y)
}
\arguments{
\item{X}{matrix}

\item{Y}{matrix}
}
\description{
Sum of entry multiplication.
}
\examples{
#Example 1
X   <- matrix(c(1,2,3,4,5,6), ncol = 3)
Y   <- matrix(2, ncol = 3, nrow = 2)
EntryMult(X,Y)
 
}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

