% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_utils.R
\name{estimator_var_from_ci}
\alias{estimator_var_from_ci}
\title{Mean estimator's variance from asymptotically normal confidence interval}
\usage{
estimator_var_from_ci(upper_limit, lower_limit, confidence_level = 0.95)
}
\arguments{
\item{upper_limit}{Upper bound for the confidence interval}

\item{lower_limit}{Lower bound for the confidence interval}

\item{confidence_level}{Confidence level of the confidence interval}
}
\value{
The variance of the estimator
}
\description{
Auxiliary function to estimate the estimator's variance (_i.e. standard error
squared) from an asymptotically normal confidence interval of level \code{1 - alpha} by implementing
the following function:

\loadmathjax
\mjdeqn{
 \text{var} = \Bigg(\dfrac{U - L}{2\cdot Z_{1 - \alpha/2}}\Bigg)^2
}{se = (U - L)/2*qnorm(1 - alpha/2)}
where \code{U} and \code{L} are the upper and lower bounds of the interval
}
\note{
Remember that the variance of the mean-estimator from asymptotically normal
confidence intervals is the estimator's standard error squared
}
\examples{
# Calculate the confidence interval for a normal mean
x <- rnorm(100)
interval_mu_x <- t.test(x)$conf.int
estimator_var_from_ci(interval_mu_x[1], interval_mu_x[2], 0.95)

}
\seealso{
\code{\link[=estimator_se_from_ci]{estimator_se_from_ci()}}
}
