% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{get_fraction_type}
\alias{get_fraction_type}
\title{Get whether object is population attributable fraction or potential impact fraction}
\usage{
get_fraction_type(object)
}
\arguments{
\item{object}{A \code{pif_sim} object}
}
\description{
Returns whether a \code{pif_sim} object was specified by the creator as population
attributable fraction or potential impact fraction
}
\examples{
#Example 1
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pifsim <- pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)
get_fraction_type(pifsim)
}
