% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{get_theta_simulations}
\alias{get_theta_simulations}
\title{Get bootstrap simulations for \code{theta}}
\usage{
get_theta_simulations(object)
}
\arguments{
\item{object}{A \code{pif_class} object}
}
\description{
Returns the bootstrap simulations for the parameter \code{theta} of \code{pif} or \code{paf}
}
\examples{
#Example 1
data(ensanut)
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pifsim <- pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)
get_theta_simulations(pifsim)
}
\seealso{
\code{\link[=get_bootstrap_simulations]{get_bootstrap_simulations()}}
}
