% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{is_paf}
\alias{is_paf}
\title{Check if object is population attributable fraction}
\usage{
is_paf(x)
}
\arguments{
\item{x}{An object}
}
\value{
Boolean whether an object is a potential impact fraction or not.
}
\description{
Method for checking whether the object is a population attributable fraction
}
\note{
The population attributable fraction is a type of potential impact fraction so
the \code{is_pif} method will return \code{TRUE} even when estimating \code{\link[=paf]{paf()}}.
}
\examples{
#Example 1
data(ensanut)
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
pafsim <- paf(design,
  theta = log(c(1.05, 1.38)), rr,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)
is_paf(pafsim)
}
\seealso{
\code{\link[=is_pif]{is_pif()}} \code{\link[=get_fraction_type]{get_fraction_type()}}
}
