% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{is_pif}
\alias{is_pif}
\title{Check if object is potential impact fraction}
\usage{
is_pif(x)
}
\arguments{
\item{x}{An object}
}
\value{
Boolean whether an object is a potential impact fraction or not.
}
\description{
Method for checking whether the object is a potential impact fraction
}
\note{
The population attributable fraction is a type of potential impact fraction so
the \code{is_pif} method will return \code{TRUE} even when estimating \code{\link[=paf]{paf()}}.
}
\examples{
#Example 1
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pifsim <- pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)
is_pif(pifsim)
}
\seealso{
\code{\link[=is_paf]{is_paf()}} \code{\link[=get_fraction_type]{get_fraction_type()}}
}
