% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismulticonvex.R
\name{ismulticonvex}
\alias{ismulticonvex}
\title{Multivariate convexity checker}
\usage{
ismulticonvex(f, xmin, xmax, tmin, tmax, maxval = 1000, tol = 1e-08)
}
\arguments{
\item{f}{Function in which convexity will be checked. Note that \code{f} must 
be a function of two vectors \code{theta} and \code{X} with \code{f = f(X,theta)}}

\item{xmin}{Minimum plaussible value of each entry of \code{X} can be found. That is: 
\code{xmin[i]} <= \code{X[i]} <= \code{xmax[i]}}

\item{xmax}{Maximum plaussible value of each entry of \code{X} can be found. That is: 
\code{xmin[i]} <= \code{X[i]} <= \code{xmax[i]}}

\item{tmin}{Minimum plaussible value of \code{theta} can be found. That is: 
\code{tmin[i]} <= \code{theta[i]} <= \code{tmax[i]}}

\item{tmax}{Maximum plaussible value of \code{theta} can be found. That is: 
             \code{tmin[i]} <= \code{theta[i]} <= \code{tmax[i]}           
             
**Optional**}

\item{maxval}{Number of values to sample to discard a function as not being convex.}

\item{tol}{Tolerance of the method (\code{var} assumed \code{0} if \code{var < tol})}
}
\description{
Checks a function \code{f} for any indication of it not being convex. 
             **Note that passing this test does not imply convexity.**
}
\examples{
#Check function is convex
ismulticonvex(function(X,theta){X*exp(theta[1] + 0.12*theta[2])}, 0, 1, c(0,0.12), c(0.1, 1))

#Check function is convex
ismulticonvex(function(X,theta){sin(X[1]^(theta[1])) *
                cos(X[2]*theta[2])}, c(0,0), c(1,1), c(0,0), c(1, 1))

#Check function is concave. (Recall function f is concave if -f is convex)
ismulticonvex(function(X,theta){- sqrt(X[1]*X[2]*X[3]*theta[1])}, c(0,0,0), c(1,1,1), 1, 2)

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}
}

