% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{n_bootstrap}
\alias{n_bootstrap}
\title{Extract number of bootstrap simulations}
\usage{
n_bootstrap(x)
}
\arguments{
\item{x}{A \code{pif_sim} object}
}
\description{
Method for obtaining the number of bootstrap simulations used to construct
a \code{pif_sim} object.
}
\examples{
#Example 1
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pifsim <- pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)
n_bootstrap(pifsim)

}
