% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf.app.R
\name{paf.app}
\alias{paf.app}
\title{PAF approximate}
\usage{
paf.app(meanx, varx, n, beta, varbeta, ci.level = 0.95)
}
\arguments{
\item{meanx}{mean of the exposure data}

\item{varx}{variance of the exposure data}

\item{n}{sample size of exposure}

\item{beta}{beta coefficient in relative risk}

\item{varbeta}{variance of beta coefficient in relative risk}

\item{ci.level}{\% confidence interval of PAF estimate}
}
\value{
list of PAF estimate and corresponding confidence interval
}
\description{
Estimates the population aggregate fraction with the mean and
variance of the exposure data
}
\examples{
x <- rweibull(100, 1.2, 1.66)
paf.app(mean(x), var(x), length(x), log(1.27), 0.002)
}
