% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_confidence_approximate_loglinear.R
\name{paf.confidence.approximate.loglinear}
\alias{paf.confidence.approximate.loglinear}
\title{Confidence Intervals for the Population Attributable Fraction when only mean and variance of exposure values is available using loglinear method}
\usage{
paf.confidence.approximate.loglinear(Xmean, Xvar, thetahat, thetavar, rr,
  nsim = 1000, confidence = 95, check_thetas = TRUE)
}
\arguments{
\item{Xmean}{Mean value of exposure levels from a previous study.}

\item{Xvar}{Variance of exposure levels from a previous study.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetavar}{Estimator of standard deviation of thetahat (usually standard error)}

\item{rr}{Function for relative risk


**Optional**}

\item{nsim}{Number of simulations}

\item{confidence}{Confidence level \% (default 95)}

\item{check_thetas}{Check that theta parameters are correctly inputed}
}
\description{
Confidence intervals for the Population Attributable Fraction for the approximate method where only mean and variance from a previous study is available.For relative risk inyective functions, the PAF is inyective, and intervals can be calculated for log(PAF), and then transformed to PAF CI.
}
\examples{
#Example 1
set.seed(46987)
rr      <- function(X,theta){exp(X*theta)}
Xmean   <- 3
Xvar    <- 1
theta   <- 0.4
thetasd <- 0.001
paf.confidence.approximate.loglinear(Xmean, Xvar, theta, thetasd, rr)

#Example 2: Compare paf.variance.approximate with paf.variance.loglinear
X1       <- rnorm(1000,3,.5)
X2       <- rnorm(1000,4,1)
X        <- as.matrix(cbind(X1,X2))
Xmean    <- colMeans(X)
Xvar     <- cov(X)
theta    <- c(0.12, 0.17)
thetasd  <- matrix(c(0.001, 0.00001, 0.00001, 0.004), byrow = TRUE, nrow = 2)
rr       <- function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}

paf.confidence.approximate.loglinear(Xmean, Xvar, theta, thetasd, rr)
paf.confidence.loglinear(X, theta, thetasd, rr)

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

