% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf.emp.R
\name{paf.emp}
\alias{paf.emp}
\title{PAF empirical}
\usage{
paf.emp(x, beta, varbeta, ci.level = 0.95)
}
\arguments{
\item{x}{vector of exposure values}

\item{beta}{beta coefficient in relative risk}

\item{varbeta}{variance of beta coefficient in relative risk}

\item{ci.level}{\% confidence interval of PAF estimate}
}
\value{
list of PAF estimate and corresponding confidence interval
}
\description{
Estimates the population aggregate fraction with individual
level exposure data
}
\examples{
x <- rweibull(100, 1.2, 1.66)
paf.emp(x, log(1.27), 0.002)
}
