% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_exponential.R
\name{paf.exponential}
\alias{paf.exponential}
\title{Population Attributable Fraction with Exponential Relative Risk 
  Function}
\usage{
paf.exponential(X, thetahat, weights = rep(1/nrow(as.matrix(X)),
  nrow(as.matrix(X))), method = c("empirical", "kernel", "approximate"),
  Xvar = var(X), deriv.method.args = list(),
  deriv.method = c("Richardson", "complex"), adjust = 1, n = 512,
  ktype = c("gaussian", "epanechnikov", "rectangular", "triangular",
  "biweight", "cosine", "optcosine"), bw = c("SJ", "nrd0", "nrd", "ucv",
  "bcv"), check_exposure = TRUE, check_rr = TRUE, check_integrals = TRUE)
}
\arguments{
\item{X}{Random sample (\code{data.frame}) which includes exposure 
and covariates or sample mean if approximate method is selected.}

\item{thetahat}{Estimator (\code{vector}) of \code{theta} for the Relative 
Risk function \code{rr} (including intercept).

**Optional**}

\item{weights}{Normalized survey \code{weights} for the sample \code{X}.}

\item{method}{Either \code{"empirical"} (default), \code{"kernel"} or 
\code{"approximate"}.}

\item{Xvar}{Variance of exposure levels (for \code{"approximate"} 
method)}

\item{deriv.method.args}{\code{method.args} for 
\code{\link[numDeriv]{hessian}} (for \code{"approximate"} method).}

\item{deriv.method}{\code{method} for \code{\link[numDeriv]{hessian}}. 
Don't change this unless you know what you are doing (for 
\code{"approximate"} method).}

\item{adjust}{Adjust bandwith parameter from density (for \code{"kernel"} 
method) from \code{\link[stats]{density}}.}

\item{n}{Number of equally spaced points at which the density (for 
\code{"kernel"} method) is to be estimated (see 
\code{\link[stats]{density}}).}

\item{ktype}{\code{"kernel"} type:  \code{"gaussian"}, 
\code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, 
\code{"biweight"}, \code{"cosine"}, \code{"optcosine"} (for \code{kernel} 
method). Additional information on kernels in \code{\link[stats]{density}}}

\item{bw}{Smoothing bandwith parameter from density (for 
\code{"kernel"} method) from \code{\link[stats]{density}}. Default 
\code{"SJ"}.}

\item{check_exposure}{Check that exposure \code{X} is positive and numeric.}

\item{check_rr}{Check that Relative Risk function \code{rr} equals 
\code{1} when evaluated at \code{0}.}

\item{check_integrals}{Check that counterfactual of theoretical minimum risk 
exposure and relative risk's expected values are well defined for this 
scenario.}
}
\value{
paf      Estimate of Population Attributable Fraction with 
  exponential relative risk
}
\description{
Function that calculates the Population Attributable Fraction 
  \code{\link{paf}} with exponential relative risk function  \code{rr} given 
  by \deqn{RR(X; \theta) = e^{\sum \theta_i X_i}}
}
\note{
\code{"approximate"} method should be the last choice. In practice 
  \code{"empirical"} should be prefered as convergence is faster than 
  \code{"kernel"} for most functions. In addition, the scope of 
  \code{"kernel"} is limited as it does not work with multivariate exposure 
  data \code{X}.

\code{\link{paf.exponential}} is a wrapper for \code{\link{paf}} with 
  exponential relative risk.
}
\examples{

#Example 1: Univariate relative risk
#----------------------------------------
set.seed(18427)
X <- rnorm(100, 3, .5)
thetahat <- 0.12
paf.exponential(X, thetahat) #Exponential risk given exp(0.12*X)

#This is the same as doing:
paf(X, thetahat, rr = function(X, theta){exp(X*theta)})

#Same example with kernel method
paf.exponential(X, thetahat,  method = "kernel")

#Same example with approximate method
Xmean <- mean(X)
Xvar  <- var(X)
paf.exponential(Xmean, thetahat, method = "approximate", Xvar = Xvar)

#Example 2: Multivariate relative risk
#----------------------------------------
X     <- matrix(c(rnorm(100,2,.7), rnorm(100,4,1)), ncol=2)
theta <- c(0.3,0.1)
paf.exponential(X,theta) #Exponential risk given exp(0.3*X1 + 0.1*X2)

}
\author{
Rodrigo Zepeda Tello \email{rzepeda17@gmail.com}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}
\references{
Vander Hoorn, S., Ezzati, M., Rodgers, A., Lopez, A. D., & 
  Murray, C. J. (2004). \emph{Estimating attributable burden of disease from 
  exposure and hazard data. Comparative quantification of health risks: 
  global and regional burden of disease attributable to selected major risk 
  factors}. Geneva: World Health Organization, 2129-40.
}
\seealso{
\code{\link{paf}} for Population Attributable Fraction (with 
  arbitrary relative risk) \code{\link{pif}} for Potential Impact Fraction 
  estimation.
  
  See \code{\link{paf.linear}} for PAF with ready-to-use linear relative risk
  function.
  
  For more information on kernels see \code{\link[stats]{density}}.
}

