% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf.R
\name{paf}
\alias{paf}
\title{Population Attributable Fraction}
\usage{
paf(X, thetahat, rr, weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))),
  method = c("empirical", "kernel"), ktype = "epanechnikov", bw = "nrd0",
  adjust = 1, npoints = 1000)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{rr}{Function for relative risk


**Optional**}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{method}{Either \code{empirical} (default) or \code{kernel}.}

\item{ktype}{\code{kernel} type from  \code{gaussian}, \code{epanechnikov}, \code{rectangular},
\code{triangular}, \code{biweight}, \code{cosine}, \code{optcosine}}

\item{bw}{Smoothing bandwith parameter from density}

\item{adjust}{Adjust bandwith parameter from density}

\item{npoints}{Number of points}
}
\value{
paf      Estimate of population attributable fraction
}
\description{
Function that calculates the population attributable fraction
}
\examples{

#Example with risk given by HR
set.seed(18427)
X <- rnorm(100)
thetahat <- 0.12
paf(X, thetahat, function(X, theta){exp(theta*X)})

#Same example with kernel method
paf(X, thetahat, function(X, theta){exp(theta*X)}, method = "kernel")

#Example with linear relative risk
paf(X, thetahat, function(X, theta){theta*X + 1})

#Multivariate example
set.seed(18427)
X1 <- rnorm(1000)
X2 <- rnorm(1000)
X  <- as.matrix(cbind(X1,X2))
thetahat <- c(0.12, 0.03)
rr <- function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}
paf(X, thetahat, rr) 

\dontrun{
#Multivariate cases cannot be evaluated with kernel method
paf(X, thetahat, rr, method = "kernel") 
}


}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

