% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_variance_approximate.R
\name{paf.variance.approximate}
\alias{paf.variance.approximate}
\title{Approximate Variance for the Population Attributable Fraction using the approximate method}
\usage{
paf.variance.approximate(Xmean, Xvar, thetahat, thetasd, rr,
  check_thetas = TRUE, nsim = 1000)
}
\arguments{
\item{Xmean}{Mean value of exposure levels from a previous study.}

\item{Xvar}{Variance of exposure levels from a previous study.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetasd}{Estimator of standard deviation of thetahat (usually standard error)}

\item{rr}{Function for relative risk


**Optional**}

\item{check_thetas}{Checks that theta parameters are correctly inputed}

\item{nsim}{Number of simulations}
}
\description{
Function that calculates approximate variance to the population attributable fraction.
}
\examples{

#Example 1
set.seed(46987)
rr      <- function(X,theta){exp(X*theta)}
Xmean   <- 3
Xvar    <- 1
theta   <- 1.2
thetasd <- 0.15
.Xmean  <- as.matrix(Xmean)
.Xvar   <- as.matrix(Xvar)
paf.variance.approximate(Xmean,Xvar,theta,thetasd,rr)

#Example 2: Compare paf.variance.approximate with paf.variance.linear
X1       <- rnorm(1000,3,.5)
X2       <- rnorm(1000,4,1)
X        <- as.matrix(cbind(X1,X2))
Xmean    <- colMeans(X)
Xvar     <- cov(X)
.Xmean   <- matrix(Xmean, ncol = length(Xmean))
.Xvar    <- matrix(Xvar, ncol = sqrt(length(Xvar)))
theta    <- c(0.12, 0.17)
thetasd  <- matrix(c(0.001, 0.00001, 0.00001, 0.004), byrow = TRUE, nrow = 2)
rr       <- function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}
paf.variance.approximate(Xmean,Xvar,theta,thetasd,rr)
paf.variance.linear(X, theta, thetasd, rr)


}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

