% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_variance_linear.R
\name{paf.variance.linear}
\alias{paf.variance.linear}
\title{Approximate Confidence Intervals for the Potential Impact Fraction}
\usage{
paf.variance.linear(X, thetahat, thetasd, rr,
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), nsim = 1000)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetasd}{Estimator of standard deviation of thetahat (usually standard error)}

\item{rr}{Function for relative risk


**Optional**}

\item{weights}{Survey \code{weights} for the random sample \code{X}}
}
\description{
Function that calculates confidence intervals to the potential impact fraction
}
\examples{

#Example with risk given by HR
set.seed(18427)
X <- rnorm(100)
thetahat <- 0.12
thetavar <- 0.1
paf.variance.linear(X, thetahat, thetavar, function(X, theta){exp(theta*X)})

#With larger sample the variance reduces
set.seed(18427)
X <- rnorm(10000)
thetahat <- 0.12
thetavar <- 0.1
paf.variance.linear(X, thetahat, thetavar, function(X, theta){exp(theta*X)})

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

