% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_utils.R
\name{paf_combine}
\alias{paf_combine}
\title{Joint PAF: Combine attributable fractions from independent and uncorrelated risk factors}
\usage{
paf_combine(paf_class_1, paf_class_2, ...)
}
\arguments{
\item{paf_class_1}{A \code{pif_class} object (\emph{i.e.} an estimation of \code{paf})}

\item{paf_class_2}{Another \code{pif_class} object (\emph{i.e.} an estimation of a different \code{paf})}

\item{...}{Additional \code{pif_class} objects (\emph{i.e.} estimations of \code{paf})}
}
\value{
A \code{paf_aggregated} object estimating the population attributable fraction for
individual-level data from multiple distinct attributable fractions.
}
\description{
A function to combine multiple attributable fractions from independent uncorrelated risk factors. The
function follows the formula from \insertCite{ezzati2003estimates;textual}{pifpaf}:
\loadmathjax
\mjdeqn{
 \text{PAF} = 1 - \prod\limits_{k} \big(1 - \text{PAF}_i)
}{paf = 1 - (1 - paf[1])*(1 - paf[2])*...*(1 - paf[n])}
}
\examples{
# Use the ensanut dataset
data(ensanut)

# EXAMPLE 1
# Setup the survey design
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr_1 <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}

rr_2 <- function(X, theta) {
  exp(-1 +
    theta[1] * X[, "age"] + X[, "systolic_blood_pressure"] )
}

rr_3 <- function(X, theta) {
  exp(-1 +
    theta[1] *X[, "systolic_blood_pressure"] )
}

paf_1 <- paf(design,
  theta = log(c(1.25, 1.68)), rr_1,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10,
)

paf_2 <- paf(design,
  theta = log(c(1.12, 1.45)), rr_2,
  additional_theta_arguments = c(0.02, 0.025), n_bootstrap_samples = 10,
)

paf_3 <- paf(design,
  theta = 0.24, rr_3,
  additional_theta_arguments = 0.01, n_bootstrap_samples = 10,
)

paf_combine(paf_1, paf_2, paf_3)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=pif]{pif()}} \code{\link[=paf]{paf()}}
}
