% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_app.R
\name{pif.app}
\alias{pif.app}
\title{PIF-APP: An app to calculate the potential impact fraction in a user friendly manner.}
\usage{
pif.app()
}
\description{
This function loads the app to calculate the Potential Impact Fraction \code{\link{pif}} in a friendly manner.
 One can calculate point estimates for \code{\link{pif}} and \code{\link{paf}}, confidence intervals for PAF.  
 Plots are also outputs such as sensitivity analysis, a heatmap for the counterfactual, a graphical display of the counterfactual,
 and a graph of PAF and PIF for different theta values. Most common relative risk functions and counterfactual functions are offered to the user. 
 The user can load data in .csv files.
}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}

