% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_approximate.R
\name{pif.approximate}
\alias{pif.approximate}
\title{Point Estimate of the Potential Impact Fraction via the Approximate method.}
\usage{
pif.approximate(Xmean, Xvar, thetahat, rr, cft = function(Xmean) {    
  matrix(0, ncol = ncol(as.matrix(Xmean)), nrow = nrow(as.matrix(Xmean))) })
}
\arguments{
\item{Xmean}{Mean value of exposure levels from a previous study.}

\item{Xvar}{Variance of the exposure levels.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function.}

\item{rr}{Function for relative risk.


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for \code{PAF} where counterfactual is 0 exposure.}
}
\description{
Function that calculates the potential impact fraction via the approximate method.
}
\examples{
#Example 1
Xmean     <- 2
thetahat  <- 0.12
Xvar      <- 0.2
rr        <- function(X,theta){exp(X*theta)}
cft       <- function(X){0.5*X}
pif.approximate(Xmean, Xvar, thetahat, rr, cft)
pif.approximate(Xmean, Xvar, thetahat, rr)

#Example 2: Multivariate
#Multivariate example

X1        <- 2
X2        <- 1.1
Xmean     <- as.matrix(cbind(X1,X2))
Xvar      <- matrix(c(1,.4,.4,1),ncol = 2, byrow = TRUE)
Cft       <- function(X){.25*X}
thetahat  <- c(0.12, 0.03)
rr <- function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}
pif.approximate(Xmean, Xvar, thetahat, rr) 
pif.approximate(Xmean, Xvar, thetahat, rr, Cft)

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

