% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_combine.R
\name{pif.combine}
\alias{pif.combine}
\title{Combine point estimates of PIF of same exposure for exclusive subpopulations}
\usage{
pif.combine(pif_vector, paf_vector, proportions)
}
\arguments{
\item{pif_vector}{Vector of the \code{pif} for each specific subpopulation.}

\item{paf_vector}{Vector of the \code{paf} for each specific subpopulation.}

\item{proportions}{Vector establishing the proportion of individuals in each subpopulation of \code{paf_vector}.}
}
\value{
An overall point-estimate of \code{pif} combining all subpopulations.
}
\description{
Function for fast-computing an overall PIF from subpopulation PIFs and PAFs
}
\note{
To combine \code{pif}s both \code{pif}s and \code{paf}s are required.
}
\examples{

#Example 1
#-------------
pif   <- c("Men" = 0.23, "Women" = 0.19)
paf   <- c("Men" = 0.47, "Women" = 0.51)
props <- c("Men" = 0.45, "Women" = 0.55)
pif.combine(pif, paf, props)  #total PAF


}
\author{
Rodrigo Zepeda Tello \email{rzepeda17@gmail.com}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}

