% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_conditional_variance_linear.R
\name{pif.conditional.variance.linear}
\alias{pif.conditional.variance.linear}
\title{Conditional Variance for the Potential Impact Fraction}
\usage{
pif.conditional.variance.linear(X, thetahat, rr, cft = function(Varx) {    
  matrix(0, ncol = ncol(as.matrix(Varx)), nrow = nrow(as.matrix(Varx))) },
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))))
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{rr}{Function for relative risk


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for 
the Population Attributable Fraction \code{PAF} where counterfactual is 0 exposure}

\item{weights}{Survey \code{weights} for the random sample \code{X}}
}
\description{
Function that calculates the conditional variance of the potential impact fraction (linearization).
}
\examples{

#Example with risk given by HR (PAF)
set.seed(18427)
X <- rnorm(100,3,.5)
thetahat <- 0.12
pif.conditional.variance.linear(X, thetahat,  rr = function(X, theta){exp(theta*X)})

#Example with linear counterfactual
cft      <- function(X){0.3*X}
pif.conditional.variance.linear(X, thetahat,  rr = function(X, theta){exp(theta*X)}, cft)

#Example with theta and X multivariate
set.seed(18427)
X1 <- rnorm(2000, 3,.5)
X2 <- rnorm(2000,3,.5)
X  <- as.matrix(cbind(X1,X2))
thetahat <- c(0.1, 0.03)
rr        <- function(X, theta){
           .X <- matrix(X, ncol = 2)
           exp(theta[1]*.X[,1] + theta[2]*.X[,2])
           }#' cft <- function(X){0.5*X}
pif.conditional.variance.linear(X, thetahat, rr, cft) 

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

