% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_confidence_approximate_loglinear.R
\name{pif.confidence.approximate.loglinear}
\alias{pif.confidence.approximate.loglinear}
\title{Confidence Intervals for the Potential Impact Fraction when only mean and variance of exposure values is available using loglinear method}
\usage{
pif.confidence.approximate.loglinear(Xmean, Xvar, thetahat, thetavar, rr,
  cft = function(Varx) {     matrix(0, ncol = ncol(as.matrix(Varx)), nrow =
  nrow(as.matrix(Varx))) }, nsim = 1000, confidence = 95,
  check_thetas = TRUE)
}
\arguments{
\item{Xmean}{Mean value of exposure levels from a previous study.}

\item{Xvar}{Variance of exposure levels from a previous study.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetavar}{Estimator of standard deviation of thetahat (usually standard error)}

\item{rr}{Function for relative risk


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for 
the Population Attributable Fraction \code{PAF} where counterfactual is 0 exposure}

\item{nsim}{Number of simulations}

\item{confidence}{Confidence level \% (default 95)}

\item{check_thetas}{Check that theta parameters are correctly inputed}
}
\description{
Confidence intervals for the Population Attributable Fraction for the approximate method where only mean and variance from a previous study is available.For relative risk inyective functions, the pif is inyective, and intervals can be calculated for log(pif), and then transformed to pif CI.
}
\examples{
#Example 1
set.seed(46987)
rr      <- function(X,theta){exp(X*theta)}
cft     <- function(X){0.4*X}
Xmean   <- 3
Xvar    <- 1
theta   <- 0.4
thetasd <- 0.001
pif.confidence.approximate.loglinear(Xmean, Xvar, theta, thetasd, rr)
paf.confidence.approximate.loglinear(Xmean, Xvar, theta, thetasd, rr)
pif.confidence.approximate.loglinear(Xmean, Xvar, theta, thetasd, rr, cft)
pif.confidence.approximate.loglinear(Xmean, Xvar, theta, thetasd, rr, cft = function(X){sqrt(X)})

#Example 2: Compare pif.variance.approximate with paf.variance.loglinear
X1       <- rnorm(1000,3,.5)
X2       <- rnorm(1000,4,1)
X        <- as.matrix(cbind(X1,X2))
Xmean    <- colMeans(X)
Xvar     <- cov(X)
theta    <- c(0.12, 0.17)
thetasd  <- matrix(c(0.001, 0.00001, 0.00001, 0.004), byrow = TRUE, nrow = 2)
rr       <- function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}

pif.confidence.approximate.loglinear(Xmean, Xvar, theta, thetasd, rr)
paf.confidence.loglinear(X, theta, thetasd, rr)
pif.confidence.approximate.loglinear(Xmean, Xvar, theta, thetasd, rr, cft = function(X){0.8*X})

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

