% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_confidence_approximate.R
\name{pif.confidence.approximate}
\alias{pif.confidence.approximate}
\title{Approximate Confidence Intervals for the Population Attributable Fraction}
\usage{
pif.confidence.approximate(Xmean, Xvar, thetahat, thetavar, rr,
  cft = function(Xmean) {     matrix(0, ncol = ncol(as.matrix(Xmean)), nrow =
  nrow(as.matrix(Xmean))) }, confidence = 95, nsim = 1000,
  check_thetas = TRUE)
}
\arguments{
\item{Xmean}{Mean value of exposure levels from a previous study.}

\item{Xvar}{Variance of exposure levels from a previous study.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetavar}{Estimator of standard error of thetahat (usually standard error)}

\item{rr}{Function for relative risk


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for 
the Population Attributable Fraction \code{PAF} where counterfactual is 0 exposure}

\item{confidence}{Confidence level \% (default 95)}

\item{nsim}{Number of simulations for estimation of variance}

\item{check_thetas}{Checks that theta parameters are correctly inputed}
}
\description{
Function that calculates approximate confidence intervals to the population attributable fraction
}
\examples{

#Example 1
set.seed(46987)
rr      <- function(X,theta){exp(X*theta)}
cft     <- function(X){0.5*X}
Xmean   <- 3
Xvar    <- 1
theta   <- 0.4
thetasd <- 0.001
.Xmean  <- as.matrix(Xmean)
.Xvar   <- as.matrix(Xvar)
pif.confidence.approximate(Xmean,Xvar,theta,thetasd,rr)
paf.confidence.approximate(Xmean,Xvar,theta,thetasd,rr)
pif.confidence.approximate(Xmean,Xvar,theta,thetasd,rr, cft)

#Example 2: Compare pif.variance.approximate with pif.variance.linear
X1       <- rnorm(1000,3,.5)
X2       <- rnorm(1000,4,1)
X        <- as.matrix(cbind(X1,X2))
Xmean    <- colMeans(X)
Xvar     <- cov(X)
.Xmean   <- matrix(Xmean, ncol = length(Xmean))
.Xvar    <- matrix(Xvar, ncol = sqrt(length(Xvar)))
theta    <- c(0.12, 0.17)
thetasd  <- matrix(c(0.001, 0.00001, 0.00001, 0.004), byrow = TRUE, nrow = 2)
rr       <- function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}
pif.confidence.approximate(Xmean, Xvar, theta, thetasd, rr)
paf.confidence.approximate(Xmean, Xvar, theta, thetasd, rr)
pif.confidence.approximate(Xmean, Xvar, theta, thetasd, rr, cft = function(X){sqrt(X)})
pif.confidence.approximate(Xmean, Xvar, theta, thetasd, rr, cft = function(X){0.5*X})

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

