% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_confidence_linear.R
\name{pif.confidence.linear}
\alias{pif.confidence.linear}
\title{Approximate Confidence Intervals for the Potential Impact Fraction}
\usage{
pif.confidence.linear(X, thetahat, thetavar, rr, cft = function(Varx) {    
  matrix(0, ncol = ncol(as.matrix(Varx)), nrow = nrow(as.matrix(Varx))) },
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), confidence = 95,
  nsim = 1000, check_thetas = TRUE)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetavar}{Estimator of standard error of thetahat (usually standard error)}

\item{rr}{Function for relative risk


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for 
the Population Attributable Fraction \code{PAF} where counterfactual is 0 exposure}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{confidence}{Confidence level \% (default 95)}

\item{nsim}{Number of simulations for estimation of variance}

\item{check_thetas}{Checks that theta parameters are correctly inputed}
}
\description{
Function that calculates approximate confidence intervals to the potential impact fraction.
}
\examples{

#Example with risk given by HR (PAF)
set.seed(18427)
X <- rnorm(100,3,.5)
thetahat <- 0.12
thetavar <- 0.1
pif.confidence.linear(X, thetahat, thetavar, function(X, theta){exp(theta*X)})
paf.confidence.linear(X, thetahat, thetavar, function(X, theta){exp(theta*X)})

#Example with linear counterfactual
cft      <- function(X){0.3*X}
pif.confidence.linear(X, thetahat, thetavar, function(X, theta){exp(theta*X)}, cft)

#Example with theta and X multivariate
set.seed(18427)
X1 <- rnorm(2000, 3,.5)
X2 <- rnorm(2000,3,.5)
X  <- as.matrix(cbind(X1,X2))
thetahat <- c(0.1, 0.03)
thetasd <- matrix(c(0.1, 0, 0, 0.05), byrow = TRUE, nrow = 2)
rr        <- function(X, theta){
           .X <- matrix(X, ncol = 2)
           exp(theta[1]*.X[,1] + theta[2]*.X[,2])
           }#' cft <- function(X){0.5*X}#' cft <- function(X){0.5*X}
pif.confidence.linear(X, thetahat, thetasd, rr, cft) 

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

