% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_confidence.R
\name{pif.confidence}
\alias{pif.confidence}
\title{Confidence intervals for the Potencial Impact Fraction}
\usage{
pif.confidence(X, thetahat, thetavar, rr, cft = function(Varx) {     matrix(0,
  ncol = ncol(as.matrix(Varx)), nrow = nrow(as.matrix(Varx))) },
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), nsim = 100,
  confidence = 95, confidence_method = c("loglinear", "linear",
  "bootstrap"), method = c("empirical", "kernel", "approximate"),
  Xvar = var(X), deriv.method.args = list(),
  deriv.method = c("Richardson", "complex"), adjust = 1, n = 512,
  ktype = c("gaussian", "epanechnikov", "rectangular", "triangular",
  "biweight", "cosine", "optcosine"), bw = c("SJ", "nrd0", "nrd", "ucv",
  "bcv"), check_exposure = TRUE, check_cft = TRUE, check_rr = TRUE,
  check_xvar = TRUE, check_integrals = TRUE, check_thetas = TRUE,
  is_paf = FALSE)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetavar}{Estimator of variance of thetahat}

\item{rr}{Function for Relative Risk which uses parameter 
  \code{theta}. The order of the parameters shound be \code{rr(X, theta)}.


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for 
the Population Attributable Fraction \code{PAF} where counterfactual is 0 exposure.}

\item{weights}{Survey \code{weights} for the random sample \code{X}.}

\item{nsim}{Number of simulations for estimation of variance.}

\item{confidence}{Confidence level \% (default 95)}

\item{confidence_method}{Either \code{linear}, \code{loglinear}, \code{bootstrap}}

\item{method}{Either \code{empirical} (default), \code{kernel} or 
\code{approximate}.}

\item{Xvar}{Variance of exposure levels (for \code{approximate} method)}

\item{deriv.method.args}{\code{method.args} for 
\code{\link[numDeriv]{hessian}} (for \code{approximate} method).}

\item{deriv.method}{\code{method} for \code{\link[numDeriv]{hessian}}. 
Don't change this unless you know what you are doing (for
\code{approximate} method).}

\item{adjust}{Adjust bandwith parameter from density (for \code{kernel}
method) from \code{\link[stats]{density}}.}

\item{n}{Number of equally spaced points at which the density (for
\code{kernel} method) is to be estimated (see
\code{\link[stats]{density}}).}

\item{ktype}{\code{kernel} type:  \code{"gaussian"}, 
\code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, 
\code{"biweight"}, \code{"cosine"}, \code{"optcosine"} (for \code{kernel}
method). Additional information on kernels in \code{\link[stats]{density}}}

\item{bw}{Smoothing bandwith parameter from density (for \code{kernel}
method) from \code{\link[stats]{density}}. Default \code{"SJ"}.}

\item{check_exposure}{Check that exposure \code{X} is positive and numeric}

\item{check_cft}{Check if counterfactual function \code{cft} reduces exposure.}

\item{check_rr}{Check that Relative Risk function \code{rr} equals 
\code{1} when evaluated at \code{0}}

\item{check_xvar}{Check if it is covariance matrix.}

\item{check_integrals}{Check that counterfactual and relative risk's expected
values are well defined for this scenario}

\item{check_thetas}{Check that theta parameters are correctly inputed}

\item{is_paf}{Boolean forcing evaluation of paf}
}
\description{
Confidence intervals for the potencial impact Fraction
}
\examples{
# Example 1: univariate example
set.seed(82392)
X         <- rnorm(1000,3,.5)
thetahat  <- 0.5
thetavar  <- 0.1
rr        <- function(X,theta){exp(theta*X)}
cft <- function(X){0.5*X}
# Default
pif.confidence(X = X, thetahat = thetahat, thetavar = thetavar, 
rr = rr, cft = cft)

# Kernel
pif.confidence(X = X, thetahat = thetahat, thetavar = thetavar, 
rr = rr, cft = cft, method = "kernel",confidence_method = "bootstrap")

 # Approximate 
 Xmean <- mean(X)
 Xvar  <- var(X)
 pif.confidence(X = Xmean, thetahat = thetahat, thetavar = thetavar, 
 rr = rr, cft = cft, method = "approximate", Xvar = Xvar)
 
 # Example 2: multivariate example
 
X1 <- rnorm(100, 3,.5)
X2 <- rnorm(100,3,.5)
X  <- as.matrix(cbind(X1,X2))
thetahat <- c(0.1, 0.03)
thetavar <- matrix(c(0.1, 0, 0, 0.05), byrow = TRUE, nrow = 2)
rr       <- function(X, theta){
  .X <- matrix(X, ncol = 2)
  exp(theta[1]*.X[,1] + theta[2]*.X[,2])
}
cft <- function(X){0.5*X}

# Default
pif.confidence(X = X, thetahat = thetahat, thetavar = thetavar, 
rr = rr, cft = cft)

 # Approximate 
 Xmean <- t(as.matrix(colMeans(X)))
 Xvar  <- cov(X)
 pif.confidence(X = Xmean, thetahat = thetahat, thetavar = thetavar, 
 rr = rr, cft = cft, method = "approximate", Xvar = Xvar)

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}

