% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_counterfactual_3D.R
\name{pif.counterfactual.3D}
\alias{pif.counterfactual.3D}
\title{3D Sensitivity Analysis of PIF with affine counterfactual}
\usage{
pif.counterfactual.3D(X, thetahat, rr, weights = rep(1/nrow(as.matrix(X)),
  nrow(as.matrix(X))), mina = 0, maxa = 1, minb = 0, maxb = mean(X),
  title = "Potential Impact Fraction (PIF) with counterfactual \\n f(X)= aX+b",
  xlab = "a", ylab = "b", zlab = "PIF")
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{rr}{Function for relative risk


**Optional**}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{mina}{Minimum for \code{a} for the counterfactual}

\item{maxa}{Maximum for \code{a} for the counterfactual}

\item{minb}{Minimum for \code{b} for the counterfactual}

\item{maxb}{Maximum for \code{b} for the counterfactual}

\item{title}{Title for the plot}

\item{xlab}{Label for the X-axis of the plot (corresponding to "a")}

\item{ylab}{Label for the Y-axis of the plot (corresponding to "b")}

\item{zlab}{Label for the Z-axis of the plot (corresponding to "PIF")}
}
\description{
Provides a graphic sensitivity analysis by varying the parameters of the counterfactual \code{a*X + b}
}
\examples{

#Example 1
X  <- rnorm(25,3)                        #Sample
rr <- function(X,theta){exp(X*theta)}    #Relative risk
theta <- 0.01                            #Estimate of theta
pif.counterfactual.3D(X, theta = theta, rr = rr)

#Example 2
X     <- rbeta(100, 1, 0.2)
theta <- c(0.12, 1)
rr    <- function(X,theta){X*theta[1] + theta[2]}
pif.counterfactual.3D(X, theta = theta, rr = rr)


}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

