% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_counterfactual_heatmap.R
\name{pif.counterfactual.heatmap}
\alias{pif.counterfactual.heatmap}
\title{Sensitivity Analysis of PIF with counterfactual}
\usage{
pif.counterfactual.heatmap(X, thetahat, rr,
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))),
  method = "empirical", legendtitle = "PIF", mina = 0.01, maxa = 0.99,
  minb = -1, maxb = 0, nmesh = 10,
  title = "Potential Impact Fraction (PIF) with counterfactual \\n f(X)= aX+b",
  xlab = "a", ylab = "b", cft = function(X, a, b) {     a * X + b },
  palette = heat.colors(nmesh))
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{rr}{Function for relative risk


**Optional**}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{mina}{Minimum for \code{a} for the counterfactual}

\item{maxa}{Maximum for \code{a} for the counterfactual}

\item{minb}{Minimum for \code{b} for the counterfactual}

\item{maxb}{Maximum for \code{b} for the counterfactual}

\item{nmesh}{Number of points in mesh (default \code{30})}

\item{title}{Title for the plot}

\item{xlab}{Label for the X-axis of the plot (corresponding to "a")}

\item{ylab}{Label for the Y-axis of the plot (corresponding to "b")}

\item{cft}{Counterfactual function with parameters \code{a} and \code{b}. Default \code{aX + b}}
}
\description{
Provides a graphic sensitivity analysis by varying the parameters of a bivariate counterfactual
}
\examples{

#Example 1
X  <- rnorm(25,3)                        #Sample
rr <- function(X,theta){exp(X*theta)}    #Relative risk
theta <- 0.01                            #Estimate of theta
pif.counterfactual.heatmap(X, theta = theta, rr = rr)

#Example 2
X     <- rbeta(100, 1, 0.2)
theta <- c(0.12, 1)
rr    <- function(X,theta){X*theta[1] + theta[2]}
cft   <- function(X, a, b){sin(a*X)*b}
pif.counterfactual.heatmap(X, theta = theta, rr = rr, cft = cft, 
     nmesh = 15, palette = rainbow(30), method = "kernel",
     title = "PIF with counterfactual cft(X) = sin(a*X)*b")

#You can also plot univariate counterfactuals
X     <- rgamma(100, 1, 0.2)
theta <- c(0.12, 1)
rr    <- function(X,theta){X*theta[1] + theta[2]}
cft   <- function(X, a, b){sqrt(a*X)}   #Leave two variables in it
pif.counterfactual.heatmap(X, theta = theta, rr = rr, 
cft = cft, minb = 0, maxb = 0, title ="Univariate counterfactual", ylab = "")

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

