% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_empirical.R
\name{pif.empirical}
\alias{pif.empirical}
\title{Point Estimate of the Potential Impact Fraction via the Empirical Method}
\usage{
pif.empirical(X, thetahat, rr, cft = function(Varx) {     matrix(0, ncol =
  ncol(as.matrix(Varx)), nrow = nrow(as.matrix(Varx))) },
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))))
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{rr}{Function for relative risk


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for \code{PAF} where counterfactual is 0 exposure}

\item{weights}{Survey \code{weights} for the random sample \code{X}}
}
\description{
Function that calculates the potential impact fraction via the empirical method
}
\examples{

#Example with risk given by HR
set.seed(18427)
X <- rnorm(100,3,.5)
thetahat <- 0.12
pif.empirical(X, thetahat, function(X, theta){exp(theta*X)})

#Same example considering counterfactual of halfing exposure
pif.empirical(X, thetahat, function(X, theta){exp(theta*X)}, cft = function(X){ 0.5*X })

#Example with linear relative risk
pif.empirical(X, thetahat, function(X, theta){theta*X + 1}, cft = function(X){ 0.5*X })

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

