% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_kernel.R
\name{pif.kernel}
\alias{pif.kernel}
\title{Kernel-based Potential Impact Fraction}
\usage{
pif.kernel(X, thetahat, rr, cft = function(Varx) {     matrix(0, ncol =
  ncol(as.matrix(Varx)), nrow = nrow(as.matrix(Varx))) },
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))),
  kernel = "epanechnikov", bw = "nrd0", adjust = 1, npoints = 1000)
}
\arguments{
\item{X}{Random sample (vector) of exposure}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{rr}{Function for relative risk


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for \code{PAF} where counterfactual is 0 exposure}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{kernel}{Kernel type from  "gaussian", "epanechnikov", "rectangular","triangular", "biweight","cosine", "optcosine"}

\item{bw}{Smoothing bandwith parameter from density}

\item{adjust}{Adjust bandwith parameter from density}

\item{npoints}{Number of points}
}
\description{
Function that calculates the potential impact fraction for a parameter theta based on kernel methods
}
\examples{

#Example with risk given by HR
set.seed(18427)
X <- rnorm(100)
thetahat <- 0.12
pif.kernel(X, thetahat, function(X, theta){exp(theta*X)})

#Same example considering counterfactual of halfing exposure
pif.kernel(X, thetahat, function(X, theta){exp(theta*X)}, cft = function(X){ 0.5*X })


}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

