% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_variance_approximate_linear.R
\name{pif.variance.approximate.linear}
\alias{pif.variance.approximate.linear}
\title{Approximate Variance for the Potential Impact Fraction using the approximate method}
\usage{
pif.variance.approximate.linear(Xmean, Xvar, thetahat, thetasd, rr,
  cft = function(Xmean) {     matrix(0, ncol = ncol(as.matrix(Xmean)), nrow =
  nrow(as.matrix(Xmean))) }, check_thetas = TRUE, nsim = 1000)
}
\arguments{
\item{Xmean}{Mean value of exposure levels from a previous study.}

\item{Xvar}{Variance of exposure levels from a previous study.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetasd}{Estimator of standard deviation of thetahat (usually standard error)}

\item{rr}{Function for relative risk


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for 
the Population Attributable Fraction \code{PAF} where counterfactual is 0 exposure.}

\item{check_thetas}{Checks that theta parameters are correctly inputed}

\item{nsim}{Number of simulations}
}
\description{
Function that calculates approximate variance to the potential impact fraction.
}
\examples{

#Example 1
set.seed(46987)
rr      <- function(X,theta){exp(X*theta)}
cft     <- function(X){0.5*X}
Xmean   <- 3
Xvar    <- 1
theta   <- 1.2
thetasd <- 0.15
pif.variance.approximate.linear(Xmean,Xvar,theta,thetasd,rr)
pif.variance.approximate.linear(Xmean,Xvar,theta,thetasd,rr,cft) 
pif.variance.approximate.linear(Xmean,Xvar,theta,thetasd,rr,cft = function(X){sqrt(X)}) 


#Example 2: Compare pif.variance.approximate with paf.variance.approximate
X1       <- rnorm(1000,3,.5)
X2       <- rnorm(1000,4,1)
X        <- as.matrix(cbind(X1,X2))
Xmean    <- colMeans(X)
Xvar     <- cov(X)
theta    <- c(0.12, 0.17)
thetasd  <- matrix(c(0.001, 0.00001, 0.00001, 0.004), byrow = TRUE, nrow = 2)
rr       <- function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}
pif.variance.approximate.linear(Xmean,Xvar,theta,thetasd,rr)
paf.variance.approximate (Xmean,Xvar,theta,thetasd,rr)
pif.variance.approximate.linear(Xmean,Xvar,theta,thetasd,rr, cft)

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

