% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{pif_class}
\alias{pif_class}
\title{pif_class class}
\usage{
pif_class(
  n_boot_simulations = class_missing,
  is_paf = class_missing,
  bootstrap_design = class_missing,
  theta_simulations = class_missing,
  pif_classulations = class_missing
)
}
\arguments{
\item{n_boot_simulations}{Number of bootstrap simulations}

\item{is_paf}{(\code{boolean}) Whether the function being estimated is the Population
Attributable Fraction (\code{is_paf = TRUE}) or the Potential Impact Fraction
(\code{is_paf = FALSE})}

\item{bootstrap_design}{The survey bootstrap design utilized for estimating \code{pif} or \code{paf}.}

\item{theta_simulations}{Simulations for the parameter \code{theta} of the relative risk.}

\item{pif_classulations}{Bootstrap simulations for the impact fraction or the attributable
fraction.}
}
\description{
An \code{\link[S7:S7_class]{S7::S7_class()}} for saving the bootstrap simulations of the potential
impact fraction \code{\link[=pif]{pif()}} and the population attributable fraction \code{\link[=paf]{paf()}}.
}
\note{
The use of this class are for development only.
}
