% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_utils.R
\name{pif_combine}
\alias{pif_combine}
\title{Joint PIF: Combine impact fractions from independent and uncorrelated risk factors}
\usage{
pif_combine(pif_class_1, pif_class_2, ...)
}
\arguments{
\item{pif_class_1}{A \code{pif_class} object (\emph{i.e.} an estimation of \code{pif} or \code{paf})}

\item{pif_class_2}{Another \code{pif_class} object (\emph{i.e.} an estimation of a different \code{pif} or \code{paf})}

\item{...}{Additional \code{pif_class} objects (\emph{i.e.} estimations of \code{pif} or \code{paf})}
}
\value{
A \code{pif_aggregated} object estimating the potential impact fraction for
individual-level data from multiple distinct impact fractions.
}
\description{
A function to combine multiple  impact fractions from independent uncorrelated risk factors. The
function follows the formula from \insertCite{ezzati2003estimates;textual}{pifpaf}:
\loadmathjax
\mjdeqn{
 \text{PIF} = 1 - \prod\limits_{k} \big(1 - \text{PIF}_i)
}{pif = 1 - (1 - pif[1])*(1 - pif[2])*...*(1 - pif[n])}
}
\examples{
# Use the ensanut dataset
data(ensanut)

# EXAMPLE 1
# Setup the survey design
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr_1 <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}

rr_2 <- function(X, theta) {
  exp(-1 +
    theta[1] * X[, "age"] + X[, "systolic_blood_pressure"] )
}

rr_3 <- function(X, theta) {
  exp(-1 +
    theta[1] *X[, "systolic_blood_pressure"] )
}

cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}

pif_1 <- pif(design,
  theta = log(c(1.25, 1.68)), rr_1, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)

pif_2 <- pif(design,
  theta = log(c(1.12, 1.45)), rr_2, cft,
  additional_theta_arguments = c(0.02, 0.025), n_bootstrap_samples = 10, parallel = FALSE
)

pif_3 <- pif(design,
  theta = 0.24, rr_3, cft,
  additional_theta_arguments = 0.01, n_bootstrap_samples = 10, parallel = FALSE
)

pif_combine(pif_1, pif_2, pif_3)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=pif]{pif()}} \code{\link[=paf]{paf()}}
}
