% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_survey_bootstrap.R
\name{pif_data_frame}
\alias{pif_data_frame}
\title{Pif from dataframe}
\usage{
pif_data_frame(df, theta, rr, cft, weights = NULL, is_paf = FALSE)
}
\description{
Pif from dataframe
}
\note{
In previous version of the package this function was called \code{pif.empirical}.
}
