% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{pif_sim}
\alias{pif_sim}
\title{pif_sim class}
\usage{
pif_sim(
  rr = class_missing,
  cft = class_missing,
  is_paf = class_missing,
  bootstrap_design = class_missing,
  theta_simulations = class_missing,
  pif_simulations = class_missing,
  uncertainty_interval_type = class_missing,
  confidence_level = class_missing
)
}
\arguments{
\item{rr}{(\code{function}) relative risk function with two parameters: a \code{data.frame} called
\code{X} containing the individual-level exposure and covariates, and \code{theta} (in that order).}

\item{cft}{(\code{function}) counterfactual function that takes a \code{data.frame}, \code{X}, of
individual-level exposure and covariates and returns a new \code{data.frame} of individual-level
counterfactual exposure and covariates.}

\item{is_paf}{(\code{boolean}) Whether the function being estimated is the Population
Attributable Fraction (\code{is_paf = TRUE}) or the Potential Impact Fraction
(\code{is_paf = FALSE})}

\item{bootstrap_design}{The survey bootstrap design utilized for estimating \code{pif} or \code{paf}.}

\item{theta_simulations}{Simulations for the parameter \code{theta} of the relative risk.}

\item{pif_simulations}{Bootstrap simulations for the impact fraction or the attributable
fraction.}

\item{uncertainty_interval_type}{(\code{string}) either \code{'wald'} (recommended) or \code{'percentile'} for
uncertainty intervals based upon the bootstrap's percentile or the Wald-type approximation
(see confidence interval section)}

\item{confidence_level}{(\code{double}) Confidence level for the uncertainty interval.
Defaults to \code{0.95}.}
}
\description{
An \code{\link[S7:S7_class]{S7::S7_class()}} for saving the bootstrap simulations of the potential
impact fraction \code{\link[=pif]{pif()}} and the population attributable fraction \code{\link[=paf]{paf()}}.
}
\note{
The use of this class are for development only.
}
