% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_survey_bootstrap.R
\name{pif_survey_bootstrap}
\alias{pif_survey_bootstrap}
\title{Potential Impact Fraction via bootstrap for survey data}
\usage{
pif_survey_bootstrap(
  design,
  theta,
  rr,
  cft = NULL,
  additional_theta_arguments,
  n_bootstrap_samples = NULL,
  theta_distribution = "default",
  parallel = TRUE,
  num_cores = 1,
  is_paf = FALSE,
  ...
)
}
\arguments{
\item{design}{(\code{survey.design}, \code{data.frame},\code{tibble}, or \code{svyrep.design}) survey data
structure. If data comes from a survey set the \code{design} with \code{\link[survey:svydesign]{survey::svydesign()}}. It can
also support a \code{\link[survey:svrepdesign]{survey::svrepdesign()}} design if your survey comes with replicates.
Finally, the model can also accommodate a \code{data.frame} or \code{tibble} with weights assuming
simple random sampling without replacement.}

\item{theta}{(\code{vector}/\code{double}) parameters of the relative risk function \code{rr}.}

\item{rr}{(\code{function}) relative risk function with two parameters: a \code{data.frame} called
\code{X} containing the individual-level exposure and covariates, and \code{theta} (in that order).}

\item{cft}{(\code{function}) counterfactual function that takes a \code{data.frame}, \code{X}, of
individual-level exposure and covariates and returns a new \code{data.frame} of individual-level
counterfactual exposure and covariates.}

\item{additional_theta_arguments}{any additional information on \code{theta} utilized
for obtaining bootstrap samples from the paramter. Options are:
\itemize{
\item (\code{double}) the \strong{variance} of \code{theta} if \code{theta} is
one dimensional and asymptotical normality is assumed (default).
\item (\code{vector}) the \strong{variances} of each entry of \code{theta} if \code{theta} is
n-dimensional and its entries are uncorrelated and asymptotical normality is assumed (default).
\item (\code{matrix}) the \strong{variance-covariance} matrix of \code{theta} if \code{theta} is
n-dimensional and its entries are correlated and asymptotical normality is assumed (default).
\item any list of arguments to pass via \code{\link[base:do.call]{base::do.call()}} to \code{theta_distribution} to
simulate samples from \code{theta} if \code{theta} is not assumed to be asymptotically normally
distributed.
}

\strong{Optional}}

\item{n_bootstrap_samples}{(\code{double}) number of bootstrap samples. If a \code{svyrep.design} is passed
as an argument, then \code{n_bootstrap_samples} represents the number of number of replicates in
the design.}

\item{theta_distribution}{(\code{function}) random number generator that follows
the distribution of the estimator \code{theta}. By default, \code{theta} is assumed to be asymptotically
normal and thus \code{theta_distribution} is set to \code{\link[mvtnorm:Mvnorm]{mvtnorm::rmvnorm()}} with
variance given by \code{additional_theta_arguments}. The number of simulations for the
\code{theta_distribution} function must be parametrized by a parameter of name \code{n}.}

\item{parallel}{(\code{boolean}) Flag indicating whether or not to do computations in parallel.
Default is \code{TRUE}.}

\item{num_cores}{(\code{int}) Number of cores; defaults to \code{\link[parallel:detectCores]{parallel::detectCores()}}.}

\item{is_paf}{(\code{boolean}) Whether the function being estimated is the Population
Attributable Fraction (\code{is_paf = TRUE}) or the Potential Impact Fraction
(\code{is_paf = FALSE})}

\item{...}{Additional parameters for \code{\link[svrep:as_bootstrap_design]{svrep::as_bootstrap_design()}}.}
}
\description{
Estimates the \strong{potential impact fraction}, \code{pif}, for
\strong{individual-level exposure} data (and covariates), \code{X}, from a
cross-sectional survey. Exposure is assumed to be associated with
a \strong{relative risk} function, \code{rr}, with parameter \code{theta}. A \strong{counterfactual}
scenario as a function of the exposure \code{cft} is assumed.

The \strong{potential impact fraction} is defined \insertCite{chan2023nonparametric;textual}{pifpaf}:
\loadmathjax
\mjdeqn{
 \text{PIF} = \dfrac{\mathbb{E}\Big[RR(X;\theta)\Big] -
   \mathbb{E}\Big[RR\big(\text{cft}(X);\theta\big)\Big]}{\mathbb{E}\Big[RR(X;\theta)\Big]}
}{pif = (mean(rr) - mean(cft))/mean(rr)}

where:
\itemize{
\item \mjseqn{X} denotes the individual-level matrix of exposure and covariates,
\item \mjseqn{\theta} represents additional parameters of the relative risk function,
\item \mjseqn{RR(X,\theta)} denotes the relative risk of exposure (and covariates) at level
\mjseqn{X} given parameters \mjseqn{\theta},
\item \mjseqn{cft(X)} denotes the counterfactual function applied to exposure and covariates,
\item \mjseqn{{\mathbb{E}\Big[RR(X;\theta)\Big]}} and
\mjseqn{{\mathbb{E}\Big[RR(\text{cft}(X);\theta)\Big]}} denote the population average relative
risk under current (observed) conditions and the relative risk under the counterfactual scenario.
\item \mjseqn{\text{PIF}} represents the potential impact fraction.
}
}
\section{Additional parallelization options}{

By default the function uses \link[foreach:foreach]{foreach::foreach} to parallelize creating a cluster
via \link[doParallel:registerDoParallel]{doParallel::registerDoParallel}. For finner parallelization control you can do:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#Create the cluster outside using whatever you want
myCluster <- parallel::makeCluster(3, type = "PSOCK")

#Register the cluster
doParallel::registerDoParallel(myCluster)

#Call the function using parallel = TRUE to use \%dopar\% and num_cores = NULL to avoid setting
pif(..., parallel = TRUE, num_cores = NULL)

#Stop the cluster
doParallel::stopCluster(myCluster)
}\if{html}{\out{</div>}}
}

\references{
\insertAllCited{}
}
\keyword{internal}
