% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf-package.R
\docType{package}
\name{pifpaf-package}
\alias{pifpaf}
\alias{pifpaf-package}
\title{pifpaf: Potential Impact Fraction and Population Attributable Fraction for Cross-Sectional Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Description: Uses a generalized method to estimate the Potential Impact Fraction (PIF) and the Population Attributable Fraction (PAF) from cross-sectional data. It creates point-estimates, confidence intervals, and estimates of variance. In addition, it generates plots for conducting sensitivity analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rodrigozepeda.github.io/pifpaf/}
}

}
\author{
\strong{Maintainer}: Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com}

Authors:
\itemize{
  \item Dalia Camacho-García-Formentí
  \item Francisco Reyes
}

Other contributors:
\itemize{
  \item Tonatiuh Barrientos-Gutiérrez [contributor]
  \item Ana Basto-Abreu [contributor]
  \item Ariela Braverman-Bronstein [contributor]
  \item Dèsirée Vidaña-Pérez [contributor]
  \item Frederick Cudhea [contributor]
}

}
\keyword{internal}
