% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_pif.R
\docType{package}
\name{pifpaf}
\alias{pifpaf}
\alias{pifpaf-package}
\title{pifpaf: A package for estimating the Potential Impact Fraction and the
Population Attributable Fraction from cross-sectional data}
\description{
The pifpaf package uses a generalized method to estimate the 
Potential Impact Fraction, (PIF) and the Population Attributable Fraction (PAF) 
from cross-sectional data. It creates point-estimates (\code{\link{pif}}, 
\code{\link{paf}}), confidence intervals (\code{\link{pif.confidence}}, 
\code{\link{paf.confidence}}),and estimates of variance. 
In addition it generates plots for conducting sensitivity analysis. 
This package was developed under funding by Bloomberg 
Philanthropies. Please see the package's vignette for more information:
\code{browseVignettes("pifpaf")}.
}

