% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpif.R
\name{plotpif}
\alias{plotpif}
\title{Plot of Potential Impact Fraction under different values of theta (univariate)}
\usage{
plotpif(X, thetamin, thetamax, rr, cft = function(Varx) {     matrix(0, ncol =
  ncol(as.matrix(Varx)), nrow = nrow(as.matrix(Varx))) },
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), npoints = 100)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetamin}{Minimum of theta for plot}

\item{thetamax}{Maximum of theta for plot}

\item{rr}{Function for relative risk

**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for \code{PAF} where counterfactual is 0 exposure}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{npoints}{Number of points in plot (default 100).}
}
\description{
Function that plots the PIF under different values.
}
\examples{

#Example with risk given by HR
set.seed(18427)
X <- rlnorm(100)
plotpif(X, 0, 2, function(X, theta){exp(theta*X)})

#Example with square root counterfactual
plotpif(X, 0, 2, function(X, theta){exp(theta*X)}, cft = function(X){sqrt(X)})

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}
}

