% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{print.pif_class}
\alias{print.pif_class}
\title{Print an impact or attributable fraction}
\usage{
\method{print}{pif_class}(x, n = 10000, ...)
}
\arguments{
\item{x}{A \code{pif_class} object}

\item{n}{Maximum number of simulations to print}

\item{...}{Additional arguments to pass to \link{summary.pif_class}}
}
\description{
Method for printing a potential impact fraction or a population attributable fraction
}
\examples{
#Example 1
data(ensanut)
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pifsim <- pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)
print(pifsim)
print(pifsim, n = 1)
}
