% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{quantile_bootstrap}
\alias{quantile_bootstrap}
\title{Extract quantiles of the potential impact fraction or
population attributable fraction bootstrap simulations}
\usage{
quantile_bootstrap(x, prob, ...)
}
\arguments{
\item{x}{A \code{pif_sim} object}

\item{prob}{numeric vector of probabilities with values between 0 and 1}

\item{...}{additional arguments to pass to \link[matrixStats:rowQuantiles]{matrixStats::colQuantiles}}
}
\description{
Method for obtaining the variance-covariance matrix of bootstrap simulations
of a \code{pif_sim} object (i.e. the point-estimate of the potential impact fraction or
population attributable fraction via the bootstrap).
}
\examples{
#Example 1
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pifsim <- pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)
quantile_bootstrap(pifsim, c(0.1, 0.3, 0.5))

}
