% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_ratio_approximate_confidence.R
\name{risk.ratio.approximate.confidence}
\alias{risk.ratio.approximate.confidence}
\title{Confidence intervals for the Risk Ratio Integral}
\usage{
risk.ratio.approximate.confidence(Xmean, Xvar, thetahat, thetasd, rr,
  nsim = 1000, confidence = 95, force.min = FALSE)
}
\arguments{
\item{Xmean}{Mean value of exposure levels from a previous study.}

\item{Xvar}{Variance of exposure levels from a previous study.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetasd}{Estimator of standard deviation of thetahat (usually standard error)}

\item{rr}{Function for relative risk


**Optional**}

\item{nsim}{Number of simulations}

\item{confidence}{Confidence level \% (default: 95)}

\item{force.min}{Boolean indicating whether to force the RR to have a 
minimum value of 1 instead of 0 (not recommended).}
}
\description{
Function that calculates confidence interval for the integral ∫RR(X;theta)f(x)dx
}
\examples{
set.seed(46987)
rr      <- function(X,theta){exp(X*theta)}
X       <- rnorm(100,3.2,1)
Xmean   <- 3.2
Xvar    <- 1
theta   <- 0.4
thetasd <- 0.001
.Xmean  <- as.matrix(Xmean)
.Xvar   <- as.matrix(Xvar)
risk.ratio.approximate.confidence(Xmean, Xvar, thetahat = theta, thetasd = thetasd, rr = rr)
}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

