% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_ratio_confidence.R
\name{risk.ratio.confidence}
\alias{risk.ratio.confidence}
\title{Confidence intervals for the Risk Ratio Integral}
\usage{
risk.ratio.confidence(X, thetahat, thetasd, rr,
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), nsim = 1000,
  confidence = 95, force.min = FALSE)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetasd}{Estimator of standard deviation of thetahat (usually standard error)}

\item{rr}{Function for relative risk


**Optional**}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{nsim}{Number of simulations}

\item{confidence}{Confidence level \% (default: 95)}

\item{force.min}{Boolean indicating whether to force the RR to have a 
minimum value of 1 instead of 0 (not recommended).}
}
\description{
Function that calculates confidence interval for the integral ∫RR(X;theta)f(x)dx
}
\examples{

#Example with risk given by HR
set.seed(18427)
X <- rnorm(100,3,.7)
thetahat <- 0.4
thetavar <- 0.1
risk.ratio.confidence(X, thetahat, thetavar, function(X, theta){exp(theta*X)})

#With larger sample the variance reduces
set.seed(18427)
X <- rnorm(10000,4,1)
thetahat <- 0.12
thetavar <- 0.1
risk.ratio.confidence(X, thetahat, thetavar, function(X, theta){exp(theta*X)})

#We can force RR's CI to be >= 1
risk.ratio.confidence(X, thetahat, thetavar, function(X, theta){exp(theta*X)}, force.min = TRUE)

#Example with theta and X multivariate
set.seed(18427)
X1 <- rnorm(1000,4,1)
X2 <- rnorm(1000,4,1)
X  <- as.matrix(cbind(X1,X2))
thetahat <- c(0.12, 0.03)
thetasd <- matrix(c(0.1, 0, 0, 0.4), byrow = TRUE, nrow = 2)
rr <- function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}
risk.ratio.confidence(X, thetahat, thetasd, rr) 

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

