% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{summary.pif_class}
\alias{summary.pif_class}
\title{Return potential impact fraction or population attributable fraction summary}
\usage{
\method{summary}{pif_class}(
  object,
  parm = NULL,
  level = 0.95,
  method = c("wald", "percentile"),
  ...
)
}
\arguments{
\item{object}{A \code{pif_class} object}

\item{parm}{a specification of which parameters are to be given confidence intervals,
either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{(\code{double}) Confidence level for the uncertainty interval.
Defaults to \code{0.95}.}

\item{method}{(\code{string}) either \code{'wald'} (recommended) or \code{'percentile'} for
uncertainty intervals based upon the bootstrap's percentile or the Wald-type approximation
(see confidence interval section)}

\item{...}{Additional methods to pass to \code{\link[=coef.pif_class]{coef.pif_class()}}}
}
\description{
Method for obtaining the summary of bootstrap simulations
of a \code{pif_class} object (i.e. the point-estimate of the potential impact fraction or
population attributable fraction via the bootstrap).
}
\examples{
#Example 1
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pifsim <- pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)
summary(pifsim)
}
