% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextInputRow.R
\name{textInputRow}
\alias{textInputRow}
\title{textInputRow}
\usage{
textInputRow(inputId, label, value = "")
}
\arguments{
\item{inputId}{Input that will be used to acces the value}

\item{label}{Display label}

\item{value}{Initial value}
}
\description{
This function arranges the display of Input boxes,
Source: Alex Brown, answer on 719016, "shiny 4 small textInput boxes side-by-side", Stack Overflow, January 15 '14, 
        http://stackoverflow.com/a/21132918
}

