% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{tidy.pif_class}
\alias{tidy.pif_class}
\title{Transform to tibble}
\usage{
\method{tidy}{pif_class}(x, ...)
}
\arguments{
\item{x}{A \code{pif_class} object}

\item{...}{Additional methods to pass to \code{\link[=summary.pif_class]{summary.pif_class()}}}
}
\description{
Method for transforming a potential impact fraction or a
population attributable fraction to a data.frame
}
\examples{
#Example 1
data(ensanut)
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pifsim <- pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)
tidy(pifsim)
}
