% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_functions.R
\name{validate_cft}
\alias{validate_cft}
\title{Validate the counterfactual function}
\usage{
validate_cft(cft)
}
\arguments{
\item{`rr`}{function}
}
\value{
The counterfactual function \code{cft}
}
\description{
A function that validates that the \code{cft} argument is valid
}
\keyword{internal}
