% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_functions.R
\name{validate_confidence_level}
\alias{validate_confidence_level}
\title{Validate confidence level}
\usage{
validate_confidence_level(confidence_level = 0.95)
}
\arguments{
\item{confidence_level}{(\code{double}) Confidence level for the uncertainty interval. If its between \code{0}
and \code{1} it uses it \emph{as is}. If it is between \code{1} and \code{100} it divides it by \code{100} to
transform it to a probability. Defaults to \code{0.95}.}
}
\value{
(\code{double}) A list containing a confidence level between \code{0} and \code{1}, and the values
of \code{alpha} and \code{1 - alpha/2}.
}
\description{
A function that takes the a confidence level as input. It can be inputed as:
\itemize{
\item A number between \code{0} and \code{1}.
\item A number between \code{1} and \code{100} which is then divided by 100.
\item Any other input generates an error.
}
}
\keyword{internal}
