% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_functions.R
\name{validate_is_paf}
\alias{validate_is_paf}
\title{Validate the is_paf parameter}
\usage{
validate_is_paf(is_paf)
}
\arguments{
\item{`is_paf`}{boolean determining if \code{pif} estimation is actually \code{paf}}
}
\value{
A boolean indicating whether is paf or pif
}
\description{
A function that validates that the \code{is_paf} argument is valid
}
\keyword{internal}
