% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_functions.R
\name{validate_n_boostrap_samples}
\alias{validate_n_boostrap_samples}
\title{Validate n_bootstrap_samples}
\usage{
validate_n_boostrap_samples(n_bootstrap_samples)
}
\arguments{
\item{n_bootstrap_samples}{The number of samples for the boostrap}
}
\value{
A valid \code{n_bootstrap_samples} (integer)
}
\description{
Internal function to validate the number of bootstrap samples
}
\keyword{internal}
