% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_functions.R
\name{validate_number_of_cores}
\alias{validate_number_of_cores}
\title{Validate number of cores}
\usage{
validate_number_of_cores(num_cores)
}
\arguments{
\item{num_cores}{(\code{int}) Number of cores; defaults to \code{\link[parallel:detectCores]{parallel::detectCores()}}.}
}
\value{
An integer with the number of cores to use.
}
\description{
Internal function to validate the number of cores variable in the model
}
\keyword{internal}
