% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_functions.R
\name{validate_parallel_setup}
\alias{validate_parallel_setup}
\title{Validate parallel setup}
\usage{
validate_parallel_setup(parallel, num_cores)
}
\arguments{
\item{parallel}{Boolean indicating whether to run argument in parallel}

\item{num_cores}{Number of cores to run the parallelization for}
}
\value{
A function to parallelize \code{\link[foreach:foreach]{foreach::foreach()}} either \verb{\%do\%} or \verb{\%dopar\%}.
}
\description{
Internal function to validate the parallelization and return the adecuate \%do\% operator
for foreach.
}
\keyword{internal}
