% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_functions.R
\name{validate_pifs}
\alias{validate_pifs}
\title{Validate list of potential impact fractions}
\usage{
validate_pifs(...)
}
\arguments{
\item{...}{Additional \code{pif_class} objects (\emph{i.e.} estimations of \code{pif} or \code{paf})}
}
\value{
Boolean variable indicating whether input is valid
}
\description{
Internal function to validate a list of potential impact fractions
}
\keyword{internal}
