% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_functions.R
\name{validate_return_replicates}
\alias{validate_return_replicates}
\title{Validate the return_replicates parameter}
\usage{
validate_return_replicates(return_replicates)
}
\arguments{
\item{`return_replicates`}{boolean determining if \code{pif} estimation is actually \code{paf}}
}
\value{
A boolean indicating whether to return replicate simulations or not
}
\description{
A function that validates that the \code{return_replicates} argument is valid
}
\keyword{internal}
