% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_functions.R
\name{validate_rr}
\alias{validate_rr}
\title{Validate the relative risk function}
\usage{
validate_rr(rr)
}
\arguments{
\item{`rr`}{function}
}
\value{
The relative risk function \code{rr}
}
\description{
A function that validates that the \code{rr} argument is valid
}
\keyword{internal}
