% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_functions.R
\name{validate_theta_distribution}
\alias{validate_theta_distribution}
\title{Validate the theta_distribution function}
\usage{
validate_theta_distribution(theta_distribution, additional_theta_arguments)
}
\arguments{
\item{`theta_distribution`}{the random number simulator for theta}

\item{`additional_theta_arguments`}{list of additional arguments to use in \code{theta_distribution}}
}
\value{
A random number generating function to simulate theta from.
}
\description{
A function that validates that the \code{theta_distribution} is indeed a function
and that one of its arguments is \code{n}.
}
\keyword{internal}
