% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pifpaf_class.R
\name{var_bootstrap}
\alias{var_bootstrap}
\title{Extract the variance of bootstrap simulations of potential impact fraction
or population attributable fraction}
\usage{
var_bootstrap(x, ...)
}
\arguments{
\item{x}{A \code{pif_class} object}

\item{...}{additional arguments to pass to \code{\link[stats:cor]{stats::var()}}}
}
\description{
Method for obtaining the variance of bootstrap simulations
of a \code{x} object (i.e. the point-estimate of the potential impact fraction or
population attributable fraction via the bootstrap).
}
\examples{
#Example 1
data(ensanut)
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pifsim <- pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10,
)
var_bootstrap(pifsim)
}
