% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings_shiny_1.R
\name{Warnings1}
\alias{Warnings1}
\title{Warnings 1}
\usage{
Warnings1(Error)
}
\arguments{
\item{Error}{Number of corresponding error}
}
\description{
This function has the possible warnings and output messages in  PIFApp for confidence intervals and point estimates.
}

