% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings_shiny_2.R
\name{Warnings2}
\alias{Warnings2}
\title{Warnings plot}
\usage{
Warnings2(Error)
}
\arguments{
\item{Error}{Number of corresponding error}
}
\description{
This function has the possible warnings and output messages in  PIFApp for plots.
}

