% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cft.R
\name{check.cft}
\alias{check.cft}
\title{Check counterfactual decreases exposure levels}
\usage{
check.cft(cft, X)
}
\arguments{
\item{cft}{counterfactual function}

\item{X}{sample of exposure values / or mean exposure, when only point estimates are available}
}
\description{
Function that checks whether the counterfactual function decreases the exposure values
}
\examples{
#Example 1 
cft <- function(X){0.5*X}
X   <- runif(100, 0,2)
check.cft(cft, X)


}

