% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_confidence.R
\name{check.confidence}
\alias{check.confidence}
\title{Check confidence (of confidence interval) is between zero and a
  hundred}
\usage{
check.confidence(confidence)
}
\arguments{
\item{confidence}{Confidence level desired with \code{0 < confidence <
100}}
}
\value{
TRUE if \code{confidence} is well defined between \code{0} and \code{100}
}
\description{
Function that verifies \code{confidence} level selected is > 0
  and < 100
}
\examples{
#Example 1 
confidence <- 95
check.confidence(confidence)

}
\seealso{
\code{\link{check.thetas}}, \code{\link{check.xvar}}, 
  \code{\link{check.cft}}, \code{\link{check.rr}},
  \code{\link{check.exposure}}, \code{\link{check.integrals}}
}
\author{
Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com}

Dalia Camacho-García-Formentí \email{daliaf172@gmail.com}
}
\keyword{internal}
