% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exposure.R
\name{check.exposure}
\alias{check.exposure}
\title{Check Exposure values}
\usage{
check.exposure(X)
}
\arguments{
\item{X}{Matrix of exposure that is evaluated in rr}
}
\description{
Function that verifies if all exposure values are greater or equal to zero.
}
\examples{
#Example 1 
X <- matrix(runif(500, 0,1))
check.exposure(X)

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}

