% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exposure.R
\name{check.exposure}
\alias{check.exposure}
\title{Check Exposure values}
\usage{
check.exposure(X)
}
\arguments{
\item{X}{Data frame of exposure that is evaluated in relative risk \code{rr}.}
}
\description{
Function that verifies if all exposure values are greater or equal than zero.
}
\examples{
#Example 1 
X <- matrix(runif(500, 0,1))
check.exposure(X)

}
\seealso{
\code{\link{check.confidence}}, \code{\link{check.thetas}}, 
  \code{\link{check.cft}}, \code{\link{check.xvar}}, 
  \code{\link{check.rr}}, \code{\link{check.integrals}}
}
\author{
Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com}

Dalia Camacho-García-Formentí \email{daliaf172@gmail.com}
}
\keyword{internal}
