% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_integrals.R
\name{check.integrals}
\alias{check.integrals}
\title{Check Integrals}
\usage{
check.integrals(meanrr, meancft)
}
\arguments{
\item{meanrr}{Mean of relative risk without counterfactual}

\item{meancft}{Mean of relative risk with counterfactual}
}
\value{
TRUE if as desired
}
\description{
Function for checking that the integrals of PIF make sense
}
\examples{
\dontrun{
check.integrals(0,1)
check.integrals(1,0)
check.integrals(1,-1)
}

}

