% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_integrals.R
\name{check.integrals}
\alias{check.integrals}
\title{Check Integrals}
\usage{
check.integrals(meanrr, meancft)
}
\arguments{
\item{meanrr}{Mean of relative risk \code{rr} without counterfactual}

\item{meancft}{Mean of relative risk \code{rr} with counterfactual}
}
\value{
bool   \code{TRUE} if as desired
}
\description{
Function for checking that the integrals of \code{\link{pif}} are nonnegative.
}
\examples{
check.integrals(1,0)

\dontrun{
check.integrals(0,1)
check.integrals(1,-1)
}

}
\seealso{
\code{\link{check.confidence}}, \code{\link{check.thetas}}, 
  \code{\link{check.cft}}, \code{\link{check.xvar}}, 
  \code{\link{check.rr}}, \code{\link{check.exposure}}
}
\author{
Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com}

Dalia Camacho-García-Formentí \email{daliaf172@gmail.com}
}
\keyword{internal}
