% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rr.R
\name{check.rr}
\alias{check.rr}
\title{Check Relative Risk}
\usage{
check.rr(X, thetahat, rr, tol = 1e-08)
}
\arguments{
\item{X}{Random sample (vector or matrix) which includes exposure and
covariates. or sample mean if approximate method is selected.}

\item{thetahat}{Estimator (vector or matrix) of \code{theta} for the 
Relative Risk function.}

\item{rr}{Function for Relative Risk which uses parameter 
\code{theta}. The order of the parameters shound be \code{rr(X, theta)}.}

\item{tol}{Tolerance for concluding numeric equality.}
}
\value{
boolean Indicating \code{TRUE} if relative risk is as desired.
}
\description{
Function for checking that relative risk equals 1 when evaluated in 0.
}
\examples{
check.rr(as.matrix(rnorm(100)), 1, function(X, theta){exp(X*theta)})

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}

