% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_thetas.R
\name{check.thetas}
\alias{check.thetas}
\title{Check theta parameters for confidence intervals}
\usage{
check.thetas(thetavar, thetahat, thetalow, thetaup, method)
}
\arguments{
\item{thetavar}{Variance of \code{thetahat}.}

\item{thetahat}{Point estimate of theta.}

\item{thetalow}{Lower bound of theta's CI.}

\item{thetaup}{Upper bound of theta's CI.}

\item{method}{Method of CI's of  \code{pif.confidence} (resp. \code{paf.confidence})}
}
\value{
bool     Boolean variable indicating if hypothesis are matched.
}
\description{
Function for checking that the theta parameters are correctly
  specified according to chosen \code{pif.confidence} and
  \code{paf.confidence} \code{confidence_method}.
}
\seealso{
\code{\link{check.confidence}}, \code{\link{check.xvar}}, 
  \code{\link{check.cft}}, \code{\link{check.rr}}, 
  \code{\link{check.exposure}}, \code{\link{check.integrals}}
}
\author{
Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com}

Dalia Camacho-García-Formentí \email{daliaf172@gmail.com}
}
\keyword{internal}
