% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Xvar.R
\name{check.xvar}
\alias{check.xvar}
\title{Check covariance of exposure values was defined.}
\usage{
check.xvar(Xvar)
}
\arguments{
\item{Xvar}{Input of covariance matrix of exposure values}
}
\value{
Xvar
}
\description{
Function that verifies Xvar was defined
}
\examples{

#Example 1
Xvar <- 0.2
check.xvar(Xvar)

}

