% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_paf.R
\name{internal_pif_data_frame}
\alias{internal_pif_data_frame}
\title{Pif from \code{data.frame}}
\usage{
internal_pif_data_frame(df, theta, rr, cft, weights = NULL, is_paf = FALSE)
}
\arguments{
\item{df}{A data.frame in which to run the \code{pif} function.}

\item{theta}{(\code{vector}/\code{double}) parameters of the relative risk function \code{rr}.}

\item{rr}{(\code{function}/\code{list}) a relative risk function with two parameters: a \code{data.frame} called
\code{X} containing the individual-level exposure and covariates, and \code{theta} (in that order). It can
also be a list of several relative risk functions to apply with each function being a
different modelling scenario.}

\item{cft}{(\code{function}/\code{list}) a counterfactual function that takes a \code{data.frame}, \code{X}, of
individual-level exposure and covariates and returns a new \code{data.frame} of individual-level
counterfactual exposure and covariates. It can also be a list of several counterfactual
functions to apply with each function being a different modelling scenario.}

\item{weights}{(\code{vector}) If you are not following the recommended version and use a
\code{svydesign} object for the design you can still use \code{weights} to associate
weights to your estimation. Beware that it might not give accurate estimations of the variance
nor the uncertainty intervals.}

\item{is_paf}{(\code{boolean}) Whether the function being estimated is the Population
Attributable Fraction (\code{is_paf = TRUE}) or the Potential Impact Fraction
(\code{is_paf = FALSE})}
}
\value{
A \code{vector} object containing the simulation for the \code{pif} (respectively \code{paf}),
the populational average relative risk and the populational average counterfactual.
}
\description{
Pif from \code{data.frame}
}
\note{
In previous version of the package this function was called \code{pif.empirical}.
}
\seealso{
\code{\link[=pif]{pif()}} \code{\link[=paf]{paf()}}
}
\keyword{internal}
