% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_combine.R
\name{paf.combine}
\alias{paf.combine}
\title{Combine point estimates of PAF of same exposure for exclusive subpopulations}
\usage{
paf.combine(paf_vector, proportions)
}
\arguments{
\item{paf_vector}{Vector of the \code{paf} for each specific subpopulation.}

\item{proportions}{Vector establishing the proportion of individuals in each subpopulation of \code{paf_vector}.}
}
\value{
An overall point-estimate of \code{paf} combining all subpopulations.
}
\description{
Function for fast-computing an overall PAF from subpopulation PAF
}
\examples{

#Example 1
#-------------
pafmen   <- 0.23
pafwomen <- 0.17
paf.combine(c(pafmen, pafwomen), c(0.45, 0.55))  #total PAF


}
\author{
Rodrigo Zepeda Tello \email{rzepeda17@gmail.com}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}

