% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_confidence_one_to_one.R
\name{paf.confidence.one2one}
\alias{paf.confidence.one2one}
\title{Approximate Confidence Intervals for the Population Attributable Fraction with one to one RR function}
\usage{
paf.confidence.one2one(X, thetahat, thetalow, thetaup, rr,
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), confidence = 95,
  check_thetas = TRUE)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetalow}{Lower bound of the confidence interval (vector)}

\item{thetaup}{Upper band of the confidence interval (vector)}

\item{rr}{Function for relative risk


**Optional**}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{confidence}{Confidence level \% (default: 95)}

\item{check_thetas}{Check that thetas are correctly specified}
}
\description{
Function that calculates approximate confidence intervals of the Population Attributable Fraction
considering a one to one Relative Risk
}
\examples{

#Example with risk given by HR
set.seed(18427)
X <- rnorm(1000, 1)
thetahat <- 0.4
thetalow <- 0.1
thetaup  <- 0.7
paf.confidence.one2one(X, thetahat, thetalow, thetaup, function(X, theta){exp(theta*X)})


#Example with theta and X multivariate
set.seed(18427)
X1 <- rnorm(1000)
X2 <- rnorm(1000)
X  <- as.matrix(cbind(X1,X2))
thetahat <- c(0.12, 0.03)
thetalow <- c(0.05, 0.01)
thetaup  <- c(0.25, 0.06)
rr <- function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}
paf.confidence.one2one(X, thetahat, thetalow, thetaup, rr) 

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

