% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_confidence.R
\name{paf.confidence}
\alias{paf.confidence}
\title{Confidence intervals for the Population Attributable Fraction}
\usage{
paf.confidence(X, thetahat, rr, thetavar = matrix(0, ncol = length(thetahat),
  nrow = length(thetahat)), thetalow = thetahat, thetaup = thetahat,
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), nsim = 100,
  confidence = 95, confidence_method = c("inverse", "one2one", "loglinear",
  "linear", "bootstrap"), method = c("empirical", "kernel", "approximate"),
  Xvar = var(X), deriv.method.args = list(),
  deriv.method = c("Richardson", "complex"), adjust = 1, n = 512,
  ktype = c("gaussian", "epanechnikov", "rectangular", "triangular",
  "biweight", "cosine", "optcosine"), bw = c("SJ", "nrd0", "nrd", "ucv",
  "bcv"), check_exposure = TRUE, check_cft = TRUE, check_rr = TRUE,
  check_xvar = TRUE, check_integrals = TRUE, check_thetas = TRUE,
  force.min = FALSE)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{rr}{Function for Relative Risk which uses parameter 
  \code{theta}. The order of the parameters shound be \code{rr(X, theta)}.


**Optional**}

\item{thetavar}{Estimator of variance of thetahat}

\item{thetalow}{Lower bound of the confidence interval (vector)}

\item{thetaup}{Upper band of the confidence interval (vector)}

\item{weights}{Survey \code{weights} for the random sample \code{X}.}

\item{nsim}{Number of simulations for estimation of variance.}

\item{confidence}{Confidence level \% (default 95)}

\item{confidence_method}{Either \code{inverse}, \code{one2one}, \code{linear}, \code{loglinear}, \code{bootstrap}}

\item{method}{Either \code{empirical} (default), \code{kernel} or 
\code{approximate}.}

\item{Xvar}{Variance of exposure levels (for \code{approximate} method)}

\item{deriv.method.args}{\code{method.args} for 
\code{\link[numDeriv]{hessian}} (for \code{approximate} method).}

\item{deriv.method}{\code{method} for \code{\link[numDeriv]{hessian}}. 
Don't change this unless you know what you are doing (for
\code{approximate} method).}

\item{adjust}{Adjust bandwith parameter from density (for \code{kernel}
method) from \code{\link[stats]{density}}.}

\item{n}{Number of equally spaced points at which the density (for
\code{kernel} method) is to be estimated (see
\code{\link[stats]{density}}).}

\item{ktype}{\code{kernel} type:  \code{"gaussian"}, 
\code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, 
\code{"biweight"}, \code{"cosine"}, \code{"optcosine"} (for \code{kernel}
method). Additional information on kernels in \code{\link[stats]{density}}}

\item{bw}{Smoothing bandwith parameter from density (for \code{kernel}
method) from \code{\link[stats]{density}}. Default \code{"SJ"}.}

\item{check_exposure}{Check that exposure \code{X} is positive and numeric}

\item{check_cft}{Check if counterfactual function \code{cft} reduces exposure.}

\item{check_rr}{Check that Relative Risk function \code{rr} equals 
\code{1} when evaluated at \code{0}}

\item{check_xvar}{Check if it is covariance matrix.}

\item{check_integrals}{Check that counterfactual and relative risk's expected
values are well defined for this scenario}

\item{check_thetas}{Check that theta parameters are correctly inputed}

\item{force.min}{Boolean indicating whether to force the \code{rr} to have a 
minimum value of 1 instead of 0 (not recommended).}
}
\description{
Confidence intervals for the population attributable fraction
}
\examples{
# Example 1: univariate example
set.seed(82392)
X         <- rnorm(1000,3,.5)
thetahat  <- 0.5
thetavar  <- 0.1
rr        <- function(X,theta){exp(theta*X)}
# Default
paf.confidence(X = X, thetahat = thetahat, thetavar = thetavar, rr = rr)
# One to one
paf.confidence(X = X, thetahat = thetahat, thetalow = 0.45, 
thetaup = 0.55, rr = rr, confidence_method = "one2one")
# Approximate 
Xmean <- mean(X)
Xvar  <- var(X)
paf.confidence(X = Xmean, thetahat = thetahat, thetavar = thetavar, 
rr = rr, method = "approximate", Xvar = Xvar)
 
# Example 2: multivariate example
 
X1 <- rnorm(100, 3,.5)
X2 <- rnorm(100,3,.5)
X  <- as.matrix(cbind(X1,X2))
thetahat <- c(0.1, 0.03)
thetavar <- matrix(c(0.1, 0, 0, 0.05), byrow = TRUE, nrow = 2)
rr       <- function(X, theta){
  .X <- matrix(X, ncol = 2)
  exp(theta[1]*.X[,1] + theta[2]*.X[,2])
}
cft <- function(X){0.5*X}

# Default
paf.confidence(X = X, thetahat = thetahat, thetavar = thetavar, rr = rr)

 # Approximate 
 Xmean <- t(as.matrix(colMeans(X)))
 Xvar  <- cov(X)
 paf.confidence(X = Xmean, thetahat = thetahat, thetavar = thetavar, 
 rr = rr, method = "approximate", Xvar = Xvar)
# One to one
paf.confidence(X = X, thetahat = thetahat, thetalow = c(0.05, 0), 
thetaup = c(0.15, 0.08), rr = rr, confidence_method = "one2one")

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}

