% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_confidence.R
\name{paf.confidence}
\alias{paf.confidence}
\title{Confidence Intervals for the Population Attributable Fraction}
\usage{
paf.confidence(X, thetahat, thetavar = NA, rr,
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), thetamin = NA,
  thetamax = NA, nsim = 1000, confidence = 95, method = c("inverse",
  "log", "linear", "one2one"), force.min = FALSE)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetavar}{Estimator of standard deviation of thetahat (usually standard error)}

\item{rr}{Function for relative risk


**Optional**}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{thetamin}{Minimum value of theta (needed if \code{method} = one2one)}

\item{thetamax}{Maximum value of theta (needed if \code{method} = one2one)}

\item{nsim}{Number of simulations}

\item{confidence}{Confidence level \% (default 95)}

\item{method}{Method of estimation of confidence intervals "inverse", "log", "linear", "one2one"
(default: inverse)}

\item{force.min}{Boolean indicating whether to force the PAF to have a 
minimum value of 0 instead of allowing negative values (not recommended).
Only valid if \code{method} = "inverse"}
}
\description{
Function that calculates confidence intervals of the Population Attributable Fraction
}
\examples{

#Example with risk given by HR
set.seed(18427)
X <- rnorm(100,4,1)
thetahat <- 0.4
thetavar <- 0.1
paf.confidence(X, thetahat, thetavar, function(X, theta){exp(theta*X)})

#With larger sample the variance reduces
set.seed(18427)
X <- rnorm(10000,3,.7)
thetahat <- 0.12
thetavar <- 0.1
paf.confidence(X, thetahat, thetavar, function(X, theta){exp(theta*X)})

#We can force PAF's CI to be >= 0 
paf.confidence(X, thetahat, thetavar, function(X, theta){exp(theta*X)},
 method = "inverse", force.min = TRUE)

#Example with theta and X multivariate
set.seed(18427)
X1 <- rnorm(1000,3,.5)
X2 <- rnorm(1000,3,.6)
X  <- as.matrix(cbind(X1,X2))
thetahat <- c(0.12, 0.03)
thetasd <- matrix(c(0.1, 0, 0, 0.4), byrow = TRUE, nrow = 2)
rr <- function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}
paf.confidence(X, thetahat, thetasd, rr) 

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

