% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_survey_bootstrap.R
\name{paf}
\alias{paf}
\title{Population Attributable Fraction}
\usage{
paf(
  design,
  theta,
  rr,
  additional_theta_arguments,
  n_bootstrap_samples = NULL,
  theta_distribution = "default",
  parallel = TRUE,
  num_cores = 1,
  weights = NULL,
  ...
)
}
\arguments{
\item{design}{(\code{survey.design}, \code{data.frame},\code{tibble}, or \code{svyrep.design}) survey data
structure. If data comes from a survey set the \code{design} with \code{\link[survey:svydesign]{survey::svydesign()}}. It can
also support a \code{\link[survey:svrepdesign]{survey::svrepdesign()}} design if your survey comes with replicates.
Finally, the model can also accommodate a \code{data.frame} or \code{tibble} with weights assuming
simple random sampling without replacement.}

\item{theta}{(\code{vector}/\code{double}) parameters of the relative risk function \code{rr}.}

\item{rr}{(\code{function}) relative risk function with two parameters: a \code{data.frame} called
\code{X} containing the individual-level exposure and covariates, and \code{theta} (in that order).}

\item{additional_theta_arguments}{any additional information on \code{theta} utilized
for obtaining bootstrap samples from the paramter. Options are:
\itemize{
\item (\code{double}) the \strong{variance} of \code{theta} if \code{theta} is
one dimensional and asymptotical normality is assumed (default).
\item (\code{vector}) the \strong{variances} of each entry of \code{theta} if \code{theta} is
n-dimensional and its entries are uncorrelated and asymptotical normality is assumed (default).
\item (\code{matrix}) the \strong{variance-covariance} matrix of \code{theta} if \code{theta} is
n-dimensional and its entries are correlated and asymptotical normality is assumed (default).
\item any list of arguments to pass via \code{\link[base:do.call]{base::do.call()}} to \code{theta_distribution} to
simulate samples from \code{theta} if \code{theta} is not assumed to be asymptotically normally
distributed.
}

\strong{Optional}}

\item{n_bootstrap_samples}{(\code{double}) number of bootstrap samples. If a \code{svyrep.design} is passed
as an argument, then \code{n_bootstrap_samples} represents the number of number of replicates in
the design.}

\item{theta_distribution}{(\code{function}) random number generator that follows
the distribution of the estimator \code{theta}. By default, \code{theta} is assumed to be asymptotically
normal and thus \code{theta_distribution} is set to \code{\link[mvtnorm:Mvnorm]{mvtnorm::rmvnorm()}} with
variance given by \code{additional_theta_arguments}. The number of simulations for the
\code{theta_distribution} function must be parametrized by a parameter of name \code{n}.}

\item{parallel}{(\code{boolean}) Flag indicating whether or not to do computations in parallel.
Default is \code{TRUE}.}

\item{num_cores}{(\code{int}) Number of cores; defaults to \code{\link[parallel:detectCores]{parallel::detectCores()}}.}

\item{weights}{(\code{vector}) If you are not following the recommended version and use a
\code{svydesign} object for the design you can still use \code{weights} to associate
weights to your estimation. Beware that it might not give accurate estimations of the variance
nor the uncertainty intervals.}

\item{...}{Additional parameters for \code{\link[svrep:as_bootstrap_design]{svrep::as_bootstrap_design()}}.}
}
\description{
Estimates the \strong{population attributable fraction}, \code{paf}, for
\strong{individual-level exposure} data (and covariates), \code{X}, from a
cross-sectional survey. Exposure is assumed to be associated with
a \strong{relative risk} function, \code{rr}, with parameter \code{theta}. A \strong{counterfactual}
scenario as a function of the exposure \code{cft} is assumed.

The \strong{population attributable fraction} is defined
\insertCite{chan2023nonparametric;textual}{pifpaf} as:
\loadmathjax
\mjdeqn{
 \text{PAF} = \dfrac{\mathbb{E}\Big[RR(X;\theta)\Big] - 1}{\mathbb{E}\Big[RR(X;\theta)\Big]}
}{pif = (mean(rr) - 1)/mean(rr)}

where:
\itemize{
\item \mjseqn{X} denotes the individual-level matrix of exposure and covariates,
\item \mjseqn{\theta} represents additional parameters of the relative risk function,
\item \mjseqn{RR(X,\theta)} denotes the relative risk of exposure (and covariates) at level
\mjseqn{X} given parameters \mjseqn{\theta},
\item \mjseqn{{\mathbb{E}\Big[RR(X;\theta)\Big]}} represents the average relative risk
in the population
\item \mjseqn{1} represents the relative risk under the theoretical minimum risk scenario,
\item \mjseqn{\text{PAF}} represents the population attributable fraction
}
}
\section{Confidence Intervals}{


Confidence intervals are estimated via the two methods specified
by \insertCite{arnab2017survey;textual}{pifpaf}.
\enumerate{
\item \strong{Wald-type confidence intervals} (which are more precise) are of the form:
\mjdeqn{
\widehat{\text{PAF}} \pm t_{1 - \alpha/2}
\sqrt{\widehat{\text{Var}}_{\text{B}}\big[\widehat{\text{PIF}}\big]}
}{pif +- qt(1 - alpha/2)*sqrt(var(PAF))}
where \mjseqn{t_{1 - \alpha/2}} stands for the percent points at level \mjseqn{1 - \alpha/2}
of Student's t-distribution, and
\mjseqn{\widehat{\text{Var}}_{\text{B}}\big[\widehat{\text{PIF}}\big]}
represents the estimated variance (via bootstrap) of the potential impact fraction estimator.
\item \strong{Percentile confidence intervals} (less precise) are of the form:
\mjdeqn{
\Big[\widehat{\text{PAF}}_{\text{B},\alpha/2}, \widehat{\text{PAF}}_{\text{B},1-\alpha/2}\Big]
}{quantile(pif, c(alpha/2, 1-alpha/2))}
where \mjseqn{\widehat{\text{PAF}}_{\text{B},k}} represents the kth sample quantile
from the bootstrap simulation of the potential impact fraction estimator.
}
}

\section{Additional parallelization options}{

By default the function uses \link[foreach:foreach]{foreach::foreach} to parallelize creating a cluster
via \link[doParallel:registerDoParallel]{doParallel::registerDoParallel}. For finner parallelization control you can do:

\if{html}{\out{<div class="sourceCode">}}\preformatted{#Create the cluster outside using whatever you want
myCluster <- parallel::makeCluster(3, type = "PSOCK")

#Register the cluster
doParallel::registerDoParallel(myCluster)

#Call the function using parallel = TRUE to use \%dopar\% and num_cores = NULL to avoid setting
paf(..., parallel = TRUE, num_cores = NULL)

#Stop the cluster
doParallel::stopCluster(myCluster)
}\if{html}{\out{</div>}}
}

\examples{
# Use the ensanut dataset
data(ensanut)

# EXAMPLE 1
# Setup the survey design
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
paf(design,
  theta = log(c(1.05, 1.38)), rr,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10, parallel = FALSE
)

# EXAMPLE 2
# Now do the same but using a replicate design
options(survey.lonely.psu = "adjust")
rep_design <- svrep::as_bootstrap_design(design, replicates = 10)
paf(rep_design,
  theta = log(c(1.05, 1.38)), rr,
  additional_theta_arguments = c(0.01, 0.03)
)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=pif]{pif()}}
}
\keyword{attributable-fraction}
\keyword{paf}
