% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_variance_linear.R
\name{paf.variance.linear}
\alias{paf.variance.linear}
\title{Approximate Variance for the Population Attributable Fraction}
\usage{
paf.variance.linear(X, thetahat, thetasd, rr,
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), nsim = 1000)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{thetasd}{Estimator of standard deviation of thetahat (usually standard error)}

\item{rr}{Function for relative risk


**Optional**}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{nsim}{Number of simulations}
}
\description{
Function that calculates approximate variance to the population attributable fraction.
}
\examples{

#Example with risk given by HR
set.seed(18427)
X <- rnorm(100,3,.6)
thetahat <- 0.12
thetavar <- 0.1
paf.variance.linear(X, thetahat, thetavar, function(X, theta){exp(theta*X)})

#With larger sample the variance reduces
set.seed(18427)
X <- rnorm(10000,4,1)
thetahat <- 0.12
thetavar <- 0.1
paf.variance.linear(X, thetahat, thetavar, function(X, theta){exp(theta*X)})

#Example with theta and X multivariate
set.seed(18427)
X1 <- rnorm(1000,4,1)
X2 <- rnorm(1000,4,1)
X  <- as.matrix(cbind(X1,X2))
thetahat <- c(0.12, 0.03)
thetasd <- matrix(c(0.1, 0, 0, 0.4), byrow = TRUE, nrow = 2)
paf.variance.linear(X, thetahat, thetasd, function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}) 

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

