% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif.app.R
\name{pif.app}
\alias{pif.app}
\title{PIF approximate estimation}
\usage{
pif.app(meanx, varx, n, beta, varbeta, a = 0, b = 1, alpha = 0.05)
}
\arguments{
\item{meanx}{mean of the exposure data}

\item{varx}{variance of the exposure data}

\item{n}{sample size of exposure}

\item{beta}{beta coefficient in relative risk}

\item{varbeta}{variance of beta coefficient in relative risk}

\item{a}{the intercept (single value) for counterfactual exposure.
Must be negative.}

\item{b}{the slope (single value) for the counterfactual exposure.
Must be between 0 and 1.}

\item{alpha}{100*(1-alpha)\% confidence interval of PAF estimate}
}
\value{
list of PIF estimate and corresponding confidence interval
}
\description{
Estimates the potential impact fraction (PIF) and population
aggregate fraction (PAF) using only the mean and variance of the exposure
data. By default, the PAF is estimated if no counterfactual exposures values
are specified.
}
\examples{
# Generate exposure data
set.seed(1)
x <- rweibull(100, 1.2, 1.66)

# Estimate PAF
pif.app(mean(x), var(x), length(x), log(1.27), 0.002)

# Estimate PIF for a counterfactual exposure of a 1 unit decrease
pif.app(mean(x), var(x), length(x), log(1.27), 0.002, a = -1)

# Estimate PIF for a counterfactual exposure of 50\% decrease
pif.app(mean(x), var(x), length(x), log(1.27), 0.002, b = 0.5)
}
