% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif.app.R
\name{pif.app}
\alias{pif.app}
\title{PIF approximate estimation with mean and variance}
\usage{
pif.app(meanx, varx, n, beta, varbeta, a = 0, b = 0, alpha = 0.05)
}
\arguments{
\item{meanx}{mean of the exposure data}

\item{varx}{variance of the exposure data}

\item{n}{sample size of exposure}

\item{beta}{beta coefficient in exponential relative risk}

\item{varbeta}{variance of beta coefficient in relative risk}

\item{a}{the intercept (single value) for counterfactual exposure.
Must be negative.}

\item{b}{the slope (single value) for the counterfactual exposure.
Must be between 0 and 1.}

\item{alpha}{100*(1-alpha)\% confidence interval of PAF estimate}
}
\value{
A list of the following:
\item{\code{pif}}{point estimate of the PIF }
\item{\code{ci}}{100*(1-alpha)\% confidence interval of PIF estimate}
}
\description{
Estimates the potential impact fraction (PIF) and population
attributable fraction (PAF) using only the mean and variance of the exposure
data. Only linear counterfactual exposures of the form g(x) = a + b * x for
the PIF are supported. By default, the PAF is estimated if no counterfactual
exposures values are specified.
}
\examples{

# Estimate PAF for an exposure distribution with mean 1.41 and variance 1.18
pif.app(meanx = 1.41, varx = 1.18, n = 100, beta = log(1.27),
varbeta = 0.002)

# Estimate PIF for a counterfactual exposure of a 1 unit decrease
pif.app(meanx = 1.41, varx = 1.18, n = 100, beta = log(1.27),
varbeta = 0.002, a = -1)

# Estimate PIF for a counterfactual exposure of 50\% decrease
pif.app(meanx = 1.41, varx = 1.18, n = 100, beta = log(1.27),
varbeta = 0.002, b = 0.5)

}
\references{
Colleen E. Chan, Rodrigo Zepeda-Tello, Dalia Camacho-Garcia-Formenti,
Frederick Cudhea, Rafael Meza, Eliane Rodrigues, Donna Spiegelman, Tonatiuh
Barrientos-Gutierrez, and Xin Zhou (2022).
Nonparametric Estimation of the Potential Impact Fraction and Population
Aggregate Fraction with Individual-Level and Aggregated Data.
\url{https://arxiv.org/pdf/2207.03597.pdf}
}
