% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif.ind.R
\name{pif.ind}
\alias{pif.ind}
\title{PIF estimation with individual-level data}
\usage{
pif.ind(x, beta, varbeta, a = 0, b = 0, alpha = 0.05)
}
\arguments{
\item{x}{vector of exposure values}

\item{beta}{beta coefficient in relative risk}

\item{varbeta}{variance of beta coefficient in relative risk}

\item{a}{the intercept (single value) for counterfactual exposure.
Must be negative.}

\item{b}{the slope (single value) for the counterfactual exposure.
Must be between 0 and 1.}

\item{alpha}{100*(1-alpha)\% confidence interval of PIF estimate}
}
\value{
list of PIF estimate and corresponding confidence interval
}
\description{
Estimates the potential impact fraction ((PIF) and population
aggregate fraction (PAF) with individual-level exposure data. By default,
the PAF is estimated if no counterfactual exposures values are specified.
}
\examples{
# Generate exposure data
set.seed(1)
x <- rweibull(100, 1.2, 1.66)

# Estimate PAF
pif.ind(x, log(1.27), 0.002)

# Estimate PIF for a counterfactual exposure of a 1 unit decrease
pif.ind(x, log(1.27), 0.002, a = -1)

# Estimate PIF for a counterfactual exposure of 50\% decrease
pif.ind(x, log(1.27), 0.002, b = 0.5)
}
