% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_plot.R
\name{pif.plot}
\alias{pif.plot}
\title{Plot of Potential Impact Fraction under different values of Relative 
  Risk's parameter theta (univariate)}
\usage{
pif.plot(X, thetalow, thetaup, rr, cft = NA,
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))),
  method = c("empirical", "kernel", "approximate"), adjust = 1, n = 512,
  mpoints = 100, Xvar = var(X), deriv.method.args = list(),
  deriv.method = c("Richardson", "complex"), ktype = c("gaussian",
  "epanechnikov", "rectangular", "triangular", "biweight", "cosine",
  "optcosine"), bw = c("SJ", "nrd0", "nrd", "ucv", "bcv"),
  confidence_method = c("bootstrap", "linear", "loglinear"),
  confidence = 95, colors = c("deepskyblue", "gray25"), xlab = "Theta",
  ylab = "PIF",
  title = "Potential Impact Fraction (PIF) under different values of theta",
  check_exposure = TRUE, check_rr = TRUE, check_integrals = TRUE,
  is_paf = FALSE)
}
\arguments{
\item{X}{Random sample (\code{data.frame}) which includes exposure 
and covariates.}

\item{thetalow}{Minimum of \code{theta} (parameter of relative risk 
\code{rr}) for plot.}

\item{thetaup}{Maximum of \code{theta} (parameter of relative risk 
\code{rr}) for plot.}

\item{rr}{\code{function} for Relative Risk which uses parameter 
\code{theta}. The order of the parameters shound be \code{rr(X, theta)}.

**Optional**}

\item{cft}{\code{function} \code{cft(X)} for counterfactual. Leave 
empty for the Population Attributable Fraction \code{\link{paf}} where 
counterfactual is 0 exposure.}

\item{weights}{Normalized survey \code{weights} for the sample \code{X}.}

\item{method}{Either \code{"empirical"} (default), \code{"kernel"} or 
\code{"approximate"}.}

\item{adjust}{Adjust bandwith parameter from density (for \code{"kernel"} 
method) from \code{\link[stats]{density}}.}

\item{n}{Number of equally spaced points at which the density (for 
\code{"kernel"} method) is to be estimated (see 
\code{\link[stats]{density}}).}

\item{mpoints}{Number of points in plot.}

\item{Xvar}{Variance of exposure levels (for \code{"approximate"} 
method)}

\item{deriv.method.args}{\code{method.args} for 
\code{\link[numDeriv]{hessian}} (for \code{"approximate"} method).}

\item{deriv.method}{\code{method} for \code{\link[numDeriv]{hessian}}. 
Don't change this unless you know what you are doing (for 
\code{"approximate"} method).}

\item{ktype}{\code{"kernel"} type:  \code{"gaussian"}, 
\code{"epanechnikov"}, \code{"rectangular"}, \code{"triangular"}, 
\code{"biweight"}, \code{"cosine"}, \code{"optcosine"} (for \code{"kernel"}
method). Additional information on kernels in \code{\link[stats]{density}}}

\item{bw}{Smoothing bandwith parameter from density (for 
\code{"kernel"} method) from \code{\link[stats]{density}}. Default 
\code{"SJ"}.}

\item{confidence_method}{Either \code{"linear"}, \code{"loglinear"}, 
\code{"bootstrap"}}

\item{confidence}{Confidence level \% (default \code{95})}

\item{colors}{Colours of plot.}

\item{xlab}{Label of x-axis in plot.}

\item{ylab}{Label of y-axis in plot.}

\item{title}{Title of plot.}

\item{check_exposure}{Check that exposure \code{X} is positive and numeric.}

\item{check_rr}{Check that Relative Risk function \code{rr} equals 
\code{1} when evaluated at \code{0}.}

\item{check_integrals}{Check that counterfactual of theoretical minimum risk 
exposure and relative risk's expected values are well defined for this 
scenario.}

\item{is_paf}{Boolean forcing evaluation of \code{\link{paf}}}
}
\value{
pif.plot       \code{\link[ggplot2]{ggplot}} object with plot of 
  Potential Impact Fraction as function of \code{theta}
}
\description{
Function that plots the \code{\link{pif}} under different values
  of an univariate parameter theta of the relative risk function \code{rr} 
  which depends on the exposure \code{X} and a \code{theta} parameter 
  (\code{rr(X, theta)})
}
\examples{

#Example 1: Exponential Relative Risk empirical method
#-----------------------------------------------------
\dontrun{
set.seed(18427)
X <- rbeta(25, 4.2, 10)
pif.plot(X, thetalow = 0, thetaup = 10, rr =  function(X, theta){exp(theta*X)})

#Same example with kernel method
pif.plot(X, thetalow = 0, thetaup = 10, rr =  function(X, theta){exp(theta*X)}, method = "kernel",
title = "Kernel method example") 
 
#Same example for approximate method. Notice that approximate method has more uncertainty
Xmean <- mean(X)
Xvar  <- var(X)
pif.plot(Xmean, thetalow = 0, thetaup = 10, rr =  function(X, theta){exp(theta*X)}, 
method = "approximate", Xvar = Xvar, title = "Approximate method example")

#Example with counterfactual
pif.plot(X, thetalow = -10, thetaup = -5, rr = function(X, theta){exp(theta*X)}, 
cft = function(X){sqrt(X)})

#Example for approximate method with square root counterfactual
#Notice how the approximate represents information loss and thus the interval
#loses precision.
pif.plot(Xmean, thetalow = -10, thetaup = -5, rr = function(X, theta){exp(theta*X)},  
cft = function(X){sqrt(X)}, method = "approximate", Xvar = Xvar) 
}

}
\author{
Rodrigo Zepeda Tello \email{rzepeda17@gmail.com}

Dalia Camacho García Formentí \email{daliaf172@gmail.com}
}
\seealso{
\code{\link{pif}} for Potential Impact Fraction estimation with
  confidence intervals \code{\link{pif.confidence}}.
  
  \code{\link{paf.plot}} for same plot with Population Attributable Fraction.
}

