% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif.R, R/pif_mean.R
\name{pif}
\alias{pif}
\title{Potential Impact Fraction}
\usage{
pif(X, thetahat, rr, cft = function(Varx) {     matrix(0, ncol =
  ncol(as.matrix(Varx)), nrow = nrow(as.matrix(Varx))) },
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), eval.cvx = TRUE,
  xmin = NA, xmax = NA, tmin = NA, tmax = NA, maxval = 1000,
  tol = 1e-08)

pif(X, thetahat, rr, cft = function(Varx) {     matrix(0, ncol =
  ncol(as.matrix(Varx)), nrow = nrow(as.matrix(Varx))) },
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), eval.cvx = TRUE,
  xmin = NA, xmax = NA, tmin = NA, tmax = NA, maxval = 1000,
  tol = 1e-08)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{rr}{Function for relative risk


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for \code{PAF} where counterfactual is 0 exposure}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{eval.cvx}{Boolean for checking if relative risk function passes convexity or concavity test.}

\item{xmin}{Minimum plaussible value of each entry of \code{X} can be found. That is: 
\code{xmin[i]} <= \code{X[i]} <= \code{xmax[i]}}

\item{xmax}{Maximum plaussible value of each entry of \code{X} can be found. That is: 
\code{xmin[i]} <= \code{X[i]} <= \code{xmax[i]}}

\item{tmin}{Minimum plaussible value of \code{theta} can be found. That is: 
\code{tmin[i]} <= \code{theta[i]} <= \code{tmax[i]}}

\item{tmax}{Maximum plaussible value of \code{theta} can be found. That is: 
\code{tmin[i]} <= \code{theta[i]} <= \code{tmax[i]}}

\item{maxval}{Number of values to sample to discard a function as not being convex.}

\item{tol}{Tolerance of the method (\code{var} assumed \code{0} if \code{var < tol})}

\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{rr}{Function for relative risk


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for \code{PAF} where counterfactual is 0 exposure}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{eval.cvx}{Boolean for checking if relative risk function passes convexity or concavity test.}

\item{xmin}{Minimum plaussible value of each entry of \code{X} can be found. That is: 
\code{xmin[i]} <= \code{X[i]} <= \code{xmax[i]}}

\item{xmax}{Maximum plaussible value of each entry of \code{X} can be found. That is: 
\code{xmin[i]} <= \code{X[i]} <= \code{xmax[i]}}

\item{tmin}{Minimum plaussible value of \code{theta} can be found. That is: 
\code{tmin[i]} <= \code{theta[i]} <= \code{tmax[i]}}

\item{tmax}{Maximum plaussible value of \code{theta} can be found. That is: 
\code{tmin[i]} <= \code{theta[i]} <= \code{tmax[i]}}

\item{maxval}{Number of values to sample to discard a function as not being convex.}

\item{tol}{Tolerance of the method (\code{var} assumed \code{0} if \code{var < tol})}
}
\description{
Function that calculates the potential impact fraction via the empirical method

Function that calculates the potential impact fraction via the empirical method
}
\examples{

#Example with risk given by HR
set.seed(18427)
X <- rlnorm(100)
thetahat <- 0.12
pif(X, thetahat, function(X, theta){exp(theta*X)})

#Same example considering counterfactual of halfing exposure
pif(X, thetahat, function(X, theta){exp(theta*X)}, cft = function(X){ 0.5*X })


#Example with linear regression


#Example with  function of risk that is not convex nor concave
set.seed(523)
X <- rgamma(100, shape = 1)
thetahat <- 1.81
pif(X, thetahat, function(X, theta){X*(theta-1.82)^3},  eval.cvx = TRUE)

#Notice that deactivating the convexity checker returns results but they might be incorrect
pif(X, thetahat, function(X, theta){X*(theta-1.82)^3}, eval.cvx = FALSE)

#Notice that specifying theta's domain limits the revision of convexity to the area
#where the function actually is convex
pif(X, thetahat, function(X, theta){X*(theta-1.82)^3}, eval.cvx = TRUE, tmin = 1, tmax =1.819)



#Example with risk given by HR
set.seed(18427)
X <- rlnorm(100)
thetahat <- 0.12
pif(X, thetahat, function(X, theta){exp(theta*X)})

#Same example considering counterfactual of halfing exposure
pif(X, thetahat, function(X, theta){exp(theta*X)}, cft = function(X){ 0.5*X })


#Example with linear regression


#Example with  function of risk that is not convex nor concave
set.seed(523)
X <- rgamma(100, shape = 1)
thetahat <- 1.81
pif(X, thetahat, function(X, theta){X*(theta-1.82)^3},  eval.cvx = TRUE)

#Notice that deactivating the convexity checker returns results but they might be incorrect
pif(X, thetahat, function(X, theta){X*(theta-1.82)^3}, eval.cvx = FALSE)

#Notice that specifying theta's domain limits the revision of convexity to the area
#where the function actually is convex
pif(X, thetahat, function(X, theta){X*(theta-1.82)^3}, eval.cvx = TRUE, tmin = 1, tmax =1.819)


}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}
}

