% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif.R
\name{pif}
\alias{pif}
\title{Potential Impact Fraction}
\usage{
pif(X, thetahat, rr, cft = function(Varx) {     matrix(0, ncol =
  ncol(as.matrix(Varx)), nrow = nrow(as.matrix(Varx))) },
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))),
  method = c("empirical", "kernel", "approximate"), Xvar = var(X),
  ktype = "epanechnikov", bw = "nrd0", adjust = 1, npoints = 1000)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates. Or sample mean if approximate method is selected.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{rr}{Function for relative risk


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for 
the Population Attributable Fraction \code{PAF} where counterfactual is 0 exposure}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{method}{Either \code{empirical} (default), \code{kernel} or \code{approximate}.}

\item{Xvar}{Variance of exposure levels.}

\item{ktype}{\code{kernel} type from  \code{gaussian}, \code{epanechnikov}, \code{rectangular},
\code{triangular}, \code{biweight}, \code{cosine}, \code{optcosine}}

\item{bw}{Smoothing bandwith parameter from density}

\item{adjust}{Adjust bandwith parameter from density}

\item{npoints}{Number of points}
}
\value{
pif      Estimate of Potential Impact Fraction
}
\description{
Function that calculates the potential impact fraction
}
\examples{

#Example with risk given by HR
set.seed(18427)
X <- rnorm(100,3,1)
thetahat <- 0.12
pif(X, thetahat, function(X, theta){exp(theta*X)})

#Same example with kernel method
pif(X, thetahat, function(X, theta){exp(theta*X)}, method = "kernel")

#Same example with approximate method
pif(X, thetahat, function(X, theta){exp(theta*X)}, method = "approximate")

#Same example with approximate method calculating mean and variance previously
Xmean <- mean(X)
Xvar  <- var(X)
pif(Xmean, thetahat, function(X, theta){exp(theta*X)}, method = "approximate", Xvar = Xvar)


#Same example considering counterfactual of halfing exposure
pif(X, thetahat, function(X, theta){exp(theta*X)}, cft = function(X){ 0.5*X }, method = "empirical")

#Example with linear relative risk
pif(X, thetahat, function(X, theta){theta*X + 1}, cft = function(X){ 0.5*X })

#Multivariate example
set.seed(18427)
X1 <- rnorm(1000,4,1)
X2 <- rnorm(1000,4,1)
X  <- as.matrix(cbind(X1,X2))
thetahat <- c(0.12, 0.03)
rr <- function(X, theta){exp(theta[1]*X[,1] + theta[2]*X[,2])}
pif(X, thetahat, rr) 

#Same multivariate example for approximate method calculating mean and variance
Xmean <- matrix(colMeans(X), ncol = 2)
Xvar  <- matrix(cov(X), ncol = 2)
pif(Xmean, thetahat, rr, method = "approximate", Xvar = Xvar)


\dontrun{
#Multivariate cases cannot be evaluated with kernel method
pif(X, thetahat, rr, method = "kernel") 
}


}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

