% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pif_bootstrap.R
\name{pif}
\alias{pif}
\title{Potential Impact Fraction}
\usage{
pif(
  design,
  theta,
  rr,
  cft = NULL,
  additional_theta_arguments,
  n_bootstrap_samples = NULL,
  theta_distribution = "default",
  is_paf = FALSE,
  weights = NULL,
  .options.future = list(seed = TRUE),
  ...
)
}
\arguments{
\item{design}{(\code{survey.design}, \code{data.frame},\code{tibble}, or \code{svyrep.design}) survey data
structure. If data comes from a survey set the \code{design} with \code{\link[survey:svydesign]{survey::svydesign()}}. It can
also support a \code{\link[survey:svrepdesign]{survey::svrepdesign()}} design if your survey comes with replicates.
Finally, the model can also accommodate a \code{data.frame} or \code{tibble} with weights assuming
simple random sampling without replacement.}

\item{theta}{(\code{vector}/\code{double}) parameters of the relative risk function \code{rr}.}

\item{rr}{(\code{function}/\code{list}) a relative risk function with two parameters: a \code{data.frame} called
\code{X} containing the individual-level exposure and covariates, and \code{theta} (in that order). It can
also be a list of several relative risk functions to apply with each function being a
different modelling scenario.}

\item{cft}{(\code{function}/\code{list}) a counterfactual function that takes a \code{data.frame}, \code{X}, of
individual-level exposure and covariates and returns a new \code{data.frame} of individual-level
counterfactual exposure and covariates. It can also be a list of several counterfactual
functions to apply with each function being a different modelling scenario.}

\item{additional_theta_arguments}{any additional information on \code{theta} utilized
for obtaining bootstrap samples from the paramter. Options are:
\itemize{
\item (\code{double}) the \strong{variance} of \code{theta} if \code{theta} is
one dimensional and asymptotical normality is assumed (default).
\item (\code{vector}) the \strong{variances} of each entry of \code{theta} if \code{theta} is
n-dimensional and its entries are uncorrelated and asymptotical normality is assumed (default).
\item (\code{matrix}) the \strong{variance-covariance} matrix of \code{theta} if \code{theta} is
n-dimensional and its entries are correlated and asymptotical normality is assumed (default).
\item any list of arguments to pass via \code{\link[base:do.call]{base::do.call()}} to \code{theta_distribution} to
simulate samples from \code{theta} if \code{theta} is not assumed to be asymptotically normally
distributed.
}

\strong{Optional}}

\item{n_bootstrap_samples}{(\code{double}) number of bootstrap samples. If a \code{svyrep.design} is passed
as an argument, then \code{n_bootstrap_samples} represents the number of number of replicates in
the design.}

\item{theta_distribution}{(\code{function}) random number generator that follows
the distribution of the estimator \code{theta}. By default, \code{theta} is assumed to be asymptotically
normal and thus \code{theta_distribution} is set to \code{\link[mvtnorm:Mvnorm]{mvtnorm::rmvnorm()}} with
variance given by \code{additional_theta_arguments}. The number of simulations for the
\code{theta_distribution} function must be parametrized by a parameter of name \code{n}.}

\item{is_paf}{(\code{boolean}) Whether the function being estimated is the Population
Attributable Fraction (\code{is_paf = TRUE}) or the Potential Impact Fraction
(\code{is_paf = FALSE})}

\item{weights}{(\code{vector}) If you are not following the recommended version and use a
\code{svydesign} object for the design you can still use \code{weights} to associate
weights to your estimation. Beware that it might not give accurate estimations of the variance
nor the uncertainty intervals.}

\item{.options.future}{List of additional options for \code{\link[doFuture:grapes-dofuture-grapes]{doFuture::\%dofuture\%()}}.}

\item{...}{Additional parameters for \code{\link[svrep:as_bootstrap_design]{svrep::as_bootstrap_design()}}.}
}
\value{
A \code{\link[=pif_class]{pif_class()}} object containing the bootstrap simulations for the
potential impact fraction, the average relative risk, and the average
counterfactual if applicable.
}
\description{
Estimates the \strong{potential impact fraction}, \code{pif}, for
\strong{individual-level exposure} data (and covariates), \code{X}, from a
cross-sectional survey. Exposure is assumed to be associated with
a \strong{relative risk} function, \code{rr}, with parameter \code{theta}. A \strong{counterfactual}
scenario as a function of the exposure \code{cft} is assumed.

The \strong{potential impact fraction} is defined \insertCite{chan2023nonparametric;textual}{pifpaf}:
\loadmathjax
\mjdeqn{
 \text{PIF} = \dfrac{\mathbb{E}\Big[RR(X;\theta)\Big] -
   \mathbb{E}\Big[RR\big(\text{cft}(X);\theta\big)\Big]}{\mathbb{E}\Big[RR(X;\theta)\Big]}
}{pif = (mean(rr) - mean(cft))/mean(rr)}

where:
\itemize{
\item \mjseqn{X} denotes the individual-level matrix of exposure and covariates,
\item \mjseqn{\theta} represents additional parameters of the relative risk function,
\item \mjseqn{RR(X,\theta)} denotes the relative risk of exposure (and covariates) at level
\mjseqn{X} given parameters \mjseqn{\theta},
\item \mjseqn{cft(X)} denotes the counterfactual function applied to exposure and covariates,
\item \mjseqn{{\mathbb{E}\Big[RR(X;\theta)\Big]}} and
\mjseqn{{\mathbb{E}\Big[RR(\text{cft}(X);\theta)\Big]}} denote the population average relative
risk under current (observed) conditions and the relative risk under the counterfactual scenario.
\item \mjseqn{\text{PIF}} represents the potential impact fraction.
}
}
\section{Additional parallelization options}{

Faster computation occurs when doing parallelization which allows to use more cores in your
machine. Parallelization utilizes  the \code{\link[future:future]{future::future()}} package. For paralelization to work you
need to establish a plan (see \code{\link[future:plan]{future::plan()}} for more information). The most common
way to create parallelization in your local machine is to do:

\if{html}{\out{<div class="sourceCode">}}\preformatted{plan(multisession) #Start parallelization
pif(...)
plan(sequential) #Return to 'normal'
}\if{html}{\out{</div>}}
}

\examples{
# Use the ensanut dataset
data(ensanut)

# EXAMPLE 1
# Setup the survey design
options(survey.lonely.psu = "adjust")
design <- survey::svydesign(data = ensanut, ids = ~1, weights = ~weight, strata = ~strata)
rr <- function(X, theta) {
  exp(-2 +
    theta[1] * X[, "age"] + theta[2] * X[, "systolic_blood_pressure"] / 100)
}
cft <- function(X) {
  X[, "systolic_blood_pressure"] <- X[, "systolic_blood_pressure"] - 5
  return(X)
}
pif(design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03), n_bootstrap_samples = 10,
)

# EXAMPLE 2
# Now do the same but using a replicate design
rep_design <- svrep::as_bootstrap_design(design, replicates = 10)
pif(rep_design,
  theta = log(c(1.05, 1.38)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03)
)

# EXAMPLE 3
# Calculate two different relative risks
rr <- list(
 function(X, theta) {exp(theta[1] * X[, "systolic_blood_pressure"] / 100)},
 function(X, theta) {exp(theta[2] * X[, "systolic_blood_pressure"] / 100 + theta[3]* X[,"age"])}
)

# Calculate three counterfactual scenarios of SBP reduction from 1 to 3 mmhg
cft <- list(
function(X){X[, "systolic_blood_pressure"]  <- X[, "systolic_blood_pressure"]  - 1; return(X)},
function(X){X[, "systolic_blood_pressure"]  <- X[, "systolic_blood_pressure"]  - 2; return(X)},
function(X){X[, "systolic_blood_pressure"]  <- X[, "systolic_blood_pressure"]  - 3; return(X)}
)

pif(design,
  theta = log(c(1.05, 1.38, 1.21)), rr, cft,
  additional_theta_arguments = c(0.01, 0.03, 0.025),
  n_bootstrap_samples = 10,
)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[=paf]{paf()}} \code{\link[=plot.pif_class]{plot.pif_class()}} \code{\link[=summary.pif_class]{summary.pif_class()}}
}
\keyword{impact-fraction}
\keyword{paf}
\keyword{pif}
