% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpif.R
\name{plotpif}
\alias{plotpif}
\title{Plot of Potential Impact Fraction under different values of theta (univariate)}
\usage{
plotpif(X, thetamin = 0, thetamax = 1, rr, cft = function(Varx) {    
  matrix(0, ncol = ncol(as.matrix(Varx)), nrow = nrow(as.matrix(Varx))) },
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), npoints = 100,
  method = c("empirical", "kernel", "approximate"), Xvar = 1,
  ktype = "epanechnikov", bw = "nrd0", adjust = 1, mpoints = 1000)
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetamin}{Minimum of theta for plot}

\item{thetamax}{Maximum of theta for plot}

\item{rr}{Function for relative risk

**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for \code{PAF} where counterfactual is 0 exposure}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{npoints}{Number of points in plot (default 100).}

\item{method}{Either \code{empirical} (default), \code{kernel} or {approximate}.}

\item{Xvar}{Variance of exposure levels.}

\item{ktype}{\code{kernel} type from  \code{gaussian}, \code{epanechnikov}, \code{rectangular},
\code{triangular}, \code{biweight}, \code{cosine}, \code{optcosine}}

\item{bw}{Smoothing bandwith parameter from density}

\item{adjust}{Adjust bandwith parameter from density}

\item{mpoints}{Number of points for kernel}
}
\description{
Function that plots the PIF under different values.
}
\examples{

#Example with risk given by HR
set.seed(18427)
library(ggplot2)
X <- rnorm(100, 4.2, 1.3)
plotpif(X, thetamin = 0, thetamax = 2, function(X, theta){exp(theta*X)}) +
 ggtitle("PAF under different values of theta \\n Empirical method")

#Example with kernel method
plotpif(X, 0, 2, function(X, theta){exp(theta*X)}, method = "kernel") +
 ggtitle("PAF under different values of theta \\n Kernel method")

#Example for approximate method
Xmean <- mean(X)
Xvar  <- var(X)
plotpif(X, 0, 2, function(X, theta){exp(theta*X)}, 
method = "approximate", Xvar = Xvar) +
 ggtitle("PAF under different values of theta \\n Approximate method")

#Example with square root counterfactual
plotpif(X, 0, 2, function(X, theta){exp(theta*X)}, cft = function(X){sqrt(X)}) + 
ggtitle("PIF under different values of theta \\n square root counterfactual  \\n Empirical method")

#Example for approximate method with square root counterfactual
plotpif(X, 0, 2, function(X, theta){exp(theta*X)},  cft = function(X){sqrt(X)},
 method = "approximate", Xvar = Xvar) +
 ggtitle("PIF under different values of theta \\n square root counterfactual  \\n Approximate method")

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}

Dalia Camacho García Formentí
}

