% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivitypif.R
\name{sensitivity.pif}
\alias{sensitivity.pif}
\title{Potential Impact Fraction Sensitivity Analysis plot}
\usage{
sensitivity.pif(X, thetahat, rr, cft = function(Varx) {     matrix(0, ncol =
  ncol(as.matrix(Varx)), nrow = nrow(as.matrix(Varx))) },
  weights = rep(1/nrow(as.matrix(X)), nrow(as.matrix(X))), n = 50,
  m = 100, filename = NA,
  title = "Sensitivity Analysis for Potential Impact Fraction (PIF)")
}
\arguments{
\item{X}{Random sample (can be vector or matrix) which includes exposure and covariates.}

\item{thetahat}{Estimative of \code{theta} for the Relative Risk function}

\item{rr}{Function for relative risk


**Optional**}

\item{cft}{Function \code{cft(X)} for counterfactual. Leave empty for \code{PAF}}

\item{weights}{Survey \code{weights} for the random sample \code{X}}

\item{n}{Number of samples to include in order to conduct sensitivity analysis}

\item{m}{Limit to number of variables to remove}

\item{filename}{Name of file for saving plot}

\item{title}{Plot title}
}
\description{
Function that plots a sensitivity analysis for the potential impact fraction by checking how estimates vary when reducing the sample
}
\examples{

#Example with risk given by HR
set.seed(18427)
X        <- rnorm(1000, 4,1) 
thetahat <- 0.02
cft      <- function(X){0.5*X}
sensitivity.pif(X, thetahat, rr = function(X, theta){exp(theta*X)}, 
                m = 100, n = 20,  title = "My Sensitivity Analysis")

}
\author{
Rodrigo Zepeda Tello \email{rodrigo.zepeda@insp.mx}
}

