% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextInputRow.R
\name{textInputRow}
\alias{textInputRow}
\title{Text input row}
\usage{
textInputRow(inputId, label, value = "")
}
\arguments{
\item{inputId}{Input that will be used to acces the value}

\item{label}{Display label}

\item{value}{Initial value}
}
\description{
This function arranges the display of Input boxes,
 code comes form answers in http://stackoverflow.com/questions/20637248/shiny-4-small-textinput-boxes-side-by-side .
}

