% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_functions.R
\name{validate_survey_design}
\alias{validate_survey_design}
\title{Validate design}
\usage{
validate_survey_design(design, n_bootstrap_samples, ...)
}
\arguments{
\item{design}{A \code{svyrep.design}, \code{survey.design} or \code{survey.design2} object.}
}
\value{
A \code{svyrep.design}. If \code{design} was already a \code{svyrep.design} returns \code{design}. Else
it returns a bootstrap \code{svyrep.design} with as many as \code{n_bootstrap_samples}
}
\description{
Internal function to validate the design whether its a \code{svyrep.design} or \code{survey.design}.
}
\keyword{internal}
